[roll] Roll fuchsia [gtest_runner] use break-on-failure option
If break_on_failure is set in RunOptions, then the test suite will be run with
the --{gtest,gunit}_break_on_failure option. If a failure is
encountered, a debugger breakpoint will be inserted at the location of
the test failure and the process will be suspended while DebugAgent
attaches to the process.

Original-Fixed: b/319915727
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973475
Original-Revision: 59f79fc7138570cb669da65ac4a5c8ae9e378ce4
GitOrigin-RevId: ca6d944f0a24e7a3b3242450b8ae11f7717b17af
Change-Id: I7e3af93c63415b6d5cd339a41ba8edd4903a4630