[roll] Roll fuchsia [kernel][vm] Fix a race with the ZX_VMO_ZERO_CHILDREN signal
Due to OnZeroChildren and OnOneChild being called in a way that is not
synchronized with the actual number of children changing, i.e. they are
not called with the VMO hierarchy lock held, there is a race possible
where the ZX_VMO_ZERO_CHILDREN signal could be set even if there is,
in fact, a child.

The solution here is to synchronize, in the dispatcher, the creation of
children with the OnZeroChildren callback, to ensure the signal cannot
get 'stuck' on the wrong signal due to a race.

With this change holding the dispatcher lock during CreateChild, the
OnOneChild callback needs to be removed, with an empty implementation
in VmObjectDispatcher not being sufficient. This is because the
child observer lock needs to be held to call OnOneChild, even if it
does nothing, and we already hold the dispatcher lock. Contrast this
to OnZeroChild, where the child observer lock was acquired to call it,
and then the implementation acquires the dispatcher lock, which is a
different acquisition order and a lock ordering violation.

However, the OnOneChild callback would not have resulted in lock
ordering violations for the IoBufferDispatcher case, and it does in
fact need this signal. To work around this the equivalent information
is instead directly returned from the CreateChildReference method
instead.

Original-Bug: 135944
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/940174
Original-Revision: 041260f837fabd7e6fef8da01f59d5718461c946
GitOrigin-RevId: 6c7adfd76164d6765098e04cf41149e4eb1105e0
Change-Id: I227a65147c1e173f348b260ba35bc795f243c551