[roll] Roll fuchsia [blobfs] Include blob in corrupt error message
Include the name of the blob in the error message that gets logged when
a blob was previously found to be corrupt. The name of the blob gets
logged with the original corruption error but that can get dropped if
the device has been powered on for a long time.

Original-Bug: b/304770949
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977156
Original-Revision: c94949f64448f26d32b57a6d1a0ccf0a24bffbc9
GitOrigin-RevId: d91d4cbf26e3bf8ec6634e7e0a8ff92a26f142cb
Change-Id: I85b9359745efd8cd2fe2246b52cc24c614665a32