[roll] Roll cobalt Create a new log source enum proto.
This is the first step towards using the internal copies when building
Cobalt Core internally. The protos need to use enums, as that is how
they are defined internally. The new enum is not used yet, but I need
to create the new file so I can add it to the internal build rules,
so as to not break the internal roller in the next change.

Original-Bug: b/319167264
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/970978
Original-Revision: 8adfed2ae914ca2e50a4d7161ff4ff7715b84c0d
GitOrigin-RevId: 6f2c9db7c74a91adfcdd4425d5b740bd2780cfdb
Change-Id: Ice8099733a0c5a01c66e9a934dd7ddf0a8c33632