[roll] Roll fuchsia [netstack3] Repurpose SocketZonedIpAddr
Rename `SocketZonedIpAddr` to `StrictlyZonedAddr` and adjust the
invariants upheld by the type.

The original intent for this type was to update it to witness
`NonMappedAddr`, but in practice the places that need a zoned address
and a `NonMappedAddr` have been written to use
`ZonedAddr<SocketIpAddr<A>>` instead.

However, in a few instances, this type has organically evolved to
represent an address that is guaranteed to have a zone, when its scope
requires a zone. This change codifies this role, by renaming the type,
updating the doc comments, and removing the `From<ZonedAddr<_, _>>` impl
that circumvented the witness.

Any places where this new role is no longer relevant have stopped using
the `SocketZonedIpAddr` type and instead are using
`ZonedAddr<SpecifiedAddr<_>>`.

Original-Bug: b/42082235, b/318875810

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969698
Original-Revision: d4d762fb32fb0cda4fe1ff9502ca0a434d25ae89
GitOrigin-RevId: 6655875b288d1af217505aabde7aba43c6e84931
Change-Id: Ie96a1384eb3e90d1ae2e71668e5c67b7550ef80e