[roll] Roll fuchsia [superproject] Roll cobalt Create a new log source enum proto.
This is the first step towards using the internal copies when building
Cobalt Core internally. The protos need to use enums, as that is how
they are defined internally. The new enum is not used yet, but I need
to create the new file so I can add it to the internal build rules,
so as to not break the internal roller in the next change.

Original-Bug: b/319167264
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt/+/970978
Original-Revision: 8adfed2ae914ca2e50a4d7161ff4ff7715b84c0d
GitOrigin-RevId: 79c26523bb89e3f8ed82624446053c65553549d4
Roller-URL: https://ci.chromium.org/b/8758169149155296977
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978392
Original-Revision: a37664b2631aef970f046fd93afc3c3ba581d1e0
Change-Id: I0eb0e36da64922b932c164c4974e52df0809c3d6