[roll] Roll fuchsia [devicetree] Update devicetree boot item check
Updated logs for when there is more than 1 boot item. The current
decision is to allow for multiple devicetrees to be part of the boot
items. The one built into the kernel image is always the first one. We
only parse that one. In future we may decide to consolidate both
bootloader and fuchsia specific devicetree.

Original-Bug: b/319913179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976801
Original-Revision: b2beeb4b3772a94953d43ac071db879128d6eac4
GitOrigin-RevId: 853fac1f1855f017269740cdb3e7175ff2ca0b24
Change-Id: I83a16f0ca130de0515a50e788d318c8b8fa4d03e