[roll] Roll fuchsia [audio][test] Add pipeline error test
Previous CL created audio-core-api-error-tests, a suite to
execute test cases that are known to generate system logging
at ERROR level. This suite is allowed to run on DEBUG builds
and emulators.

Included in that set of tests was one case from the suite
audio-renderer-api-pipeline-tests, which cannot reliably
execute on non-realtime systems and thus is limited to only
release builds and non-emulators. This test case is now
flaking in CQ, so this CL moves the test case to a new suite
audio-core-api-pipeline-error-tests that does not execute on
debug builds or emulators.

Multiply: audio-core-api-error-tests
Multiply: audio-core-api-pipeline-error-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978276
Original-Revision: bc4e90b26df7cc3539d404eca30807de101b5218
GitOrigin-RevId: df1c27c0e3a93b201036d1475a9606835ea9e54f
Change-Id: Idb3b89aa6aa31ba4a6b6c56836062cf50ed7db8e