[roll] Roll fuchsia [debugdata] Fix on-target profile merging for Rust
https://reviews.llvm.org/D138846 changed the layout of the profiles by
adding new fields into the profile header and data. This caused an issue
in on-target merging in Rust profiles because debugdata uses
InstrProfData.inc shipped by the Clang toolchain. This patch adds
support of merging profiles in the old layout as a workaround to recover
the coverage drop in Rust. It also cleans up the code that was added
as a soft transition of toolchain rolls, and regenerates the profiles
that use old profile versions.

Original-Bug: 42086151
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977320
Original-Revision: aabfd693c5d747d9dcc3a1aa7f8515fe149eb5fc
GitOrigin-RevId: 70da5880e32fc47e1c52dfb4383ddb5c824ff4a7
Change-Id: Ia76d3734099e786dc61499b0f8d26ca9e426e32f