[roll] Roll fuchsia [emu] Stop reading flags template from product bundle
Currently, there is a template file that is included in the product
bundle to that is used to generate the command line flags for starting
the emulator for that product.

As things have evolved, it has turned out it is easier/better to
have a "universal" template that has conditional logic in it to
accommodate multiple cpu types and virtual device configurations, vs.
supporting the flexibility of having a template per product.

This CL stops reading the template path from the product bundle
and instead includes the template in the emulator subtool.

Follow-on CLs will remove the template path from the product bundle.

Original-Bug: b/295031514
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979192
Original-Revision: ad1b0ebd0cb79084bc8474e5590dc3d918c5d90f
GitOrigin-RevId: f517076e032194188ca1f61602a9afd272b70994
Change-Id: I86e3ce2e3ad501a68f2ff0b36773eb0ae8704734