[roll] Roll fuchsia [bt] Define the host-watcher crate
Define a Fuchsia-specific crate for watching the status of the currently
active Bluetooth host on the system. This crate uses the
fuchsia.bluetooth.sys.HostWatcher protocol to monitor changes in the
active Host. The status changes are available as a stream of events. A
client of this library will instantiate the HostWatcher, poll the
stream, and dispatch event-specific work.

Refactor HostWatcher from the Fast Pair Provider implementation.

Test: fx test host-watcher-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959499
Original-Revision: 8b9bdc30a1a7179b192d9a6b9f511445cbf4fa76
GitOrigin-RevId: 094256dda6cece677514ccf2507d3e5020cb7d56
Change-Id: I788e26965b6278661e9f6b8f56b9a4b61dbba7fc