[roll] Roll fuchsia [recovery] Don't fall back to wiping Fxfs
Previously, if the request to wipe the data volume from fshost failed,
we would fall back to manually erasing the Fxfs partition by overwriting
the superblocks.  This was sensible when Fxfs just stored mutable data,
but now Fxfs also stores blobs, so this will brick the device.

Remove the fallback logic.  This required plumbing the fxfs_blob setting
from the build into the factory_reset component, so that we would know
whether the fshost failure should be bubbled up (if fxblob is enabled)
or to continue by trying to wipe zxcrypt (if disabled).

Note that this removes the fallback logic for the Fxfs-but-not-fxblob
case too.  I didn't think it was worth also supporting that case, but if
we felt that it was important enough, we could keep it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975104
Original-Revision: b040af26e2a44b133ae480ead1f6c57683069937
GitOrigin-RevId: 61374969b6321a054a5891239566dd1e144ffc66
Change-Id: If13ad4db4d1c283315fafa1f68ea9e2f1fdaedad