[roll] Roll fuchsia [uart] AMLogic driver update.
(1) After setting the bits to reset the port (reset rx, tx, clear error)
they must be set back to zero or IRQ mode will misbehave. This could
have been done on `InitInterrupt` but seems better to complete the reset
in the same scope that is initiated. The observed behavior without
clearing this bits is that transmitted data was read as random bytes in
the host.

(2) IRQ Handler was rewritten to use rx_fifo_count which is the number
of elements in the RX fifo. Additionally this bounds the number of
operations in a single IRQ to UP to fifo characters.

(3)Enable Rx/Tx Interrupt were clearing the entire Uart Control Register
when called, and just setting the one bit. They must be initialized
from the current hardware state.

This drive was tested and stressed in a vim3.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979593
Original-Revision: ce692ef793554cc87485d85139f32f3a1eb0f542
GitOrigin-RevId: a1ad73483e51b8ae7c6eac0185aa07855c201f9f
Change-Id: Ideed12e879617b50580c41821d238a14b2ec02c0