[roll] Roll fuchsia [amlogic-display][rdma] Change "Stopping RDMA" log severity to DEBUG.
Every empty config with no layers triggers an RDMA engine stop on
the amlogic-display driver. This is totally valid.

We should suppress the logs for non-debugging environments so that
the driver won't logspam if it receives empty configs for a long time.

Test: internal product; logspam doesn't occur anymore in "fx log".
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979996
Original-Revision: 0928c392023513c16f7a5f37f680c36b5bc2cb98
GitOrigin-RevId: 6d74a2376dde53cedd0b864ff88e503006a04c6c
Change-Id: I0b661de907afc044aeac5986dc1c44fcadb0ce83