[roll] Roll fuchsia [netstack3] Move inspect_devices to DeviceIpApi
When getting ready to move `inspect_devices`, I realized that the amount
of information we were exposing from the IP layer was really lacking.
The next realization was that to expose more information with our
current inspect pattern I'd have to either create many duplicate types
to expose to bindings or expose the internal types, and in both cases
add a lot of noise to the types exported to bindings.

Instead, introduce a new pattern where we abstract the Fuchsia inspector
instead and perform all the inspecting in core. This should massively
decrease the cost of adding new inspect data, which has been a burden
and was discouraging adding more data to inspect.

Original-Bug: b/42083910
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979737
Original-Revision: 957ea13300fa4aa201f77de93ac8301f69e23468
GitOrigin-RevId: 3606ea6672047d9a3a7e74f3a795ebb81e94bd2e
Change-Id: Ib80f07e9491bf93ff200137f0a489d426a8e1b9f