[roll] Roll fuchsia [bt][avrcp] Adjust set volume value based on previous failure.
Sometimes, set volume request can get stuck in a bad state when the
controller keeps sending the request with the same volume value.
This happens because AVRCP and Fuchsia audio settings use two different
ranges for audio values. [0, 127] for AVRCP and [0.0, 1.0] for Fuchsia
audio settings.
Attempt to mitigate this bad state by remembering the requested value of the
latest failed request and adjusting the volume value when a request with
the same volume value comes in the future.

Original-Bug: b/316961553
Test: bt-a2dp-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979214
Original-Revision: 9e21d26a69c4b9c2891aa6e34028a34395895330
GitOrigin-RevId: adc5b4e7f95031bd7773fd2f8d65b8b809cc2437
Change-Id: Ic1198ac4116e5f4f0116dc2f5d5ea72b90f80cc4