[roll] Roll fuchsia [fxfs] Hold lock during bytes_ready to maintain write ordering.
Holds the lock throughout the operation to include the read and the
update in a single locking. Also reworks the error logging and error
latching to to avoid repeating entries for the same error and moves the
message handling loop into the writer module.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/978893
Original-Revision: cbe0a187b48730b098155f32b80549a3371a76b2
GitOrigin-RevId: d7e169a6e4a4b0075d7072b2b6aaf3d71e2467ae
Change-Id: I3c02f7cae903ed87b5d591d4f86437bb475fca7a