[roll] Roll fuchsia [netstack3] for_any_device_id supports type param
Update the `for_any_device_id` macro so that it can be used with a
device type parameter. This is useful in cases where the compiler is
unable to infer the device type.

While here, make the macro public so that it can be used in bindings
code.

Original-Bug: b/42051633

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979857
Original-Revision: f1a3c03fdc10843667600773f71ae5a1b8ef7871
GitOrigin-RevId: 30de2c3c01b57709381228016cf571e77635bc76
Change-Id: I6f9f43ebf2dd930a647870d40c081f6de7deda3e