[roll] Roll fuchsia [display][banjo] Support DisplayMode as a valid source for panel capabilities.
Preivously there were two ways to specify the capabilities of a
display panel: using EDID, or using DisplayParams.

As per discussed in fxbug.dev/318642950, we are going to replace
DisplayParams with DisplayMode. In order to perform the migration
smoothly, we'll do a soft transition where both DisplayParams and
DisplayMode will be supported until we finish migration of all the
display engine drivers.

This change adds another union field to Panel -- a DisplayMode
which contains all the timing information of the panel. A driver
can choose to use DisplayMode when panel_capabilities_source
is set to DISPLAY_MODE.

Test: This change itself is a no-op.
      amlogic-display: core.nelson, core.vim3 booted.
      simple-display: fx qemu --no-virtio -g
      virtio-guest: fx qemu --virtio -g
      intel-i915: core.x64 on NUC11 booted.
      fake-display: //src/ui tests

Test: tested with other changes in the chain.
Original-Bug: 318642950
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969272
Original-Revision: 9786ad96591879ece7e6ca28dadf76d22788f24b
GitOrigin-RevId: 63e8b5cdd1723d39ea3cbb5c5fd44f9f322cc8f5
Change-Id: I83c1dca5b58fcc37e9aea7a9303653e14f65f1bc