[roll] Roll fuchsia [assembly] Remove package size check merge
This feature is no longer being used, and we have no plans to use it.
This was originally used to create a budget for the base package (system
image) which was a merge of many other packages, but we removed this
budget, because nearly all of the packages were removed from the input,
and the base package is no longer something interesting to check.

Original-Bug: b/321066699
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/980422
Original-Revision: dbf9942c7d8c277d46a606da6b881daf9e829f05
GitOrigin-RevId: 25d05e2223bfcf5ad9a308f9e1db6316a0b14bb8
Change-Id: Ia0ad027547c60cbe8355a98b2f0d3bdacf971802