[roll] Roll fuchsia [components] Dynamic components receive capabilities from Dict
Components can be now created and started with a Dict that is used as
the source for `use`d capabilities.

This allows a component to dynamically route capabilities for every
execution (via CreateChildArgs.dict), and for a specific execution
(via StartChildArgs.dict).

Compared to a dynamic offers:

  * The capability source can vary per-component, especially for
    protocols. With dynamic offers, this is only possible with
    directory/service capabilities. Dicts can provide a unique
    implementation for protocols.
  * The capability source can be set for a single component
    execution. Dynamic offers persist across component restarts,
    forcing the parent to recreate the component.

When a component connects to a capability and the Dict contains an
Open value with a key that matches the capability name, it will be
provided from the Dict.

Routing from a dict is incompatible with static routes. If a capability
in a Dict is already routed statically, the Dict will be rejected.

Original-Bug: b/301973404
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/931533
Original-Revision: 4174956ebff5ef77090cd74468fe8f6af2943f45
GitOrigin-RevId: f3aebc1ab1b28f425992e768ed8ab1f50c0bcf34
Change-Id: I5099c6532fb2f4ecead604fb58184bcef10ac64b