[roll] Roll fuchsia [starnix][iptables] Parse ipt_replace entries
This CL peeks inside the entries field and attempts to create structures
that the fuchsia.net.filter API expects. It does not call the API to
commit these structures into the system, but it validates that the bytes
are set properly, and returning EINVAL otherwise.

IPTable rules were and continue to take no effect on the system.

Original-Bug: b/307908515
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946613
Original-Revision: 9791a06bd7dc2438570b1487144a0ffe609e9711
GitOrigin-RevId: ef010741cff69eac6bebac07d37c4b12efe28b66
Change-Id: I3bfe4b41f7daededada78f2d1a96e469779f4947