[roll] Roll fuchsia [aml-sdmmc] Use dedicated dispatcher for inlining FIDL calls.
This gives a small performance boost when the core sdmmc driver is
configured to use FIDL (not Banjo) to issue requests to the aml-sdmmc
driver.

Experiments show that a dedicated dispatcher gives an inlining rate
that is close to 100% (much higher than that of a default dispatcher
that is configured to disallow sync calls).

Original-Bug: b/300145353
Test: fx test aml-sdmmc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981316
Original-Revision: 247394e7b6adff273221c23468032244858a4fc5
GitOrigin-RevId: 7f6faebcac32d819a7d5a38cacc3542be26f8450
Change-Id: Id9b2c4a71a374af1857a162f315ec01debe162c8