[roll] Roll fuchsia [zircon][migration] Introduce the "deactivating" state.
Introduce the "deactivating" state to scheduler instances.  This is
functionally a sub-state added to schedulers to indicate that the
scheduler for a CPU is in the process of "deactivating", but has not
yet fully reached the deactivated state.

This sub-state is meant to be used as a piece of bookkeeping which is
used when selecting a CPU for a thread with a migration function.
Usually, we will only consider "active" CPUs during this selection
phase.  Because of the global thread-lock, it currently should be
impossible for us to ever be in a situation where:

1) A thread with a migration function (who last ran on CPU X) needs to
   become READY (unblock, un-sleep).
2) CPU X is no longer active.

After the global thread-lock has been removed, however, there will be
a small window where it is possible for a thread (call it A) with a
migration function to become READY after its last-CPU has been flagged
as inactive.  It should only be possible, however, when there is a
thread currently in the process of calling MigrateUnpinnedThreads for
that CPU.  In such a situation, the thread responsible for migrating
threads off of CPU X will discover A in X's READY queue, where it will
handle calling the migration function for A while running on X, before
moving A to the READY queue of a CPU which is active and not equal to
X.

The new introduction of the new `cpu_deactivating_` flag will allow us
to assert that whenever we encounter a thread in this situation, that
either the CPU it needs to migrate away from is either active, or
in-active, but in the process of shutting down (meaning that our
migration function will be called by the thread who is in the
process of shutting our CPU down).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/918851
Original-Revision: 018566e94403c8bf0842deb4b30055b4c6931ce9
GitOrigin-RevId: 2e066b20b4e05be1232ae5cb7831c7d745bf60a6
Change-Id: Idc7fe4e2e591a72a01c5dbe7b7aae2b5697dea84