[roll] Roll fuchsia [driver][sdk] Improvements to compat DeviceServer
This improves the ergonomics of the DeviceServer's self initialization
logic a bit. Previously it was confusing whether the async or sync init
was happening. And the callback mechanism was also confusing. This
creates a wrapper type for each sync/async inialization so it is clear
exactly what is happening.

Original-Bug: 314040286

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979517
Original-Revision: 8bc1351462356ecdab7e35ca48310a3d61752c55
GitOrigin-RevId: 69a84c7df05aac6dfc0a4425334b47de5934b655
Change-Id: I80c96f389b1ddfc70f2e6de324203509bb473652