[roll] Roll fuchsia [brcmfmac] Migrate from threads to dispatchers
The brcmfmac driver currently spawns a number of threads to perform
various work. With the transition from Banjo to FIDL these threads
create a performance problem. Threads that are not managed by the driver
runtime are not eligible for inlining FIDL calls. That means that the
calls cannot be served on the same thread as the call, instead they are
placed in a queue and another thread will service the call. This has
significant impact on performance as it introduces extra latency and
increases CPU usage.

Instead switch to using driver runtime managed threads, in the form of
dispatchers, for those threads that are involved in the data path. By
using a synchronized dispatcher that allows sync calls these dispatchers
can function in the same way as the unmanaged threads did. Instead of
providing a thread function the same function can be posted as a task to
the dispatcher and be left to run indefinitely. The combination of a
synchronized dispatcher with allowing sync calls ensures that this does
not block any other dispatchers.

Test: Manual performance testing, relevant brcmfmac unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941458
Original-Revision: b19754ef46315f242ade83f776b481c0dadb2f5e
GitOrigin-RevId: cb16c42427c8e9b3b4bc44685ae4582542c79675
Change-Id: I9dbdf84cf7b49c9a9f5e02827f6de272797bc953