[roll] Roll fuchsia [bazel-sdk] fix extension stripping in fuchsia_cc_driver
The rstrip method in bazel does not strip a string but rather strips
a set of characters that are specified by that string. This was causing
issues when we were stripping extensions because if a name ended with
any letters in the extension they would be stripped as well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982400
Original-Revision: 96e1a7dec7d0d473369127cce7920f2def0f82d2
GitOrigin-RevId: cc9c31168b84485460747841d2b227cb1521805f
Change-Id: Ic2505e6965ed62b627031b439d625f20d684ec62