[roll] Roll scudo [scudo] Add StackDepot lock to enable/disable. (#79670)
Scudo grabs all allocator locks in a pthread_atfork before the fork, and releases them after. This allows malloc to be used in a fork child of a multithreaded process, which is expressly forbidden by the standard, but very widely used. For example, Android's init uses std::string after fork when spawning services in android::init::EnterNamespaces and other places.

Any lock that is necessary to serve an allocator call must be handled this way. Otherwise there is a possibility that the lock is held during the call to fork, which results in it being held forever in the child process, and the next operation that needs it deadlocks.

GitOrigin-RevId: cd10d0c0bbdefc92d8e134362fb90af02a514cf7
Original-Revision: 8f2b20bd35697c3651fe90e43cd6092606539b11
Change-Id: Ic764e76c9692808675dbd76a5f3ed776012978b0