[roll] Roll fuchsia [kernel][gicv2] add a spinlock to protect irq configuration
In the case of gic_configure and gic_set_affinity a RMW sequence needs
to be lock in the extraordinarily rare case where two cpus tries to
fiddle with the same line. Add a spinlock to protect.

Also replaced static with anonymous namespace, which is the bulk of the
file's changes.
Code formatter picked up a few stray old changes that hadn't been
touched in a while.

Original-Bug: b/320570343
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975100
Original-Revision: 53da13c567c09ca6bc14bae4027992b755b74e80
GitOrigin-RevId: b54638dea30ce74ed034caecc014c39ee4bb27fe
Change-Id: Ifd252a6205edaa5d4472141b8a55f73439ba5675