[roll] Roll fuchsia [kernel][vm] Discardable cleanup in destructor
Moves cleanup of discardable VMOs from DeadTransition to the
destructor. This is to avoid a race with the raw back pointer in the
discardable tracker being upgraded to a refptr while the VmCowPages
is cleaning up and removing itself from discardable tracking.

Original-Bug: 322421862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982732
Original-Revision: b273ee04fb048708fff8431037b76c72127e8dd3
GitOrigin-RevId: 70c6b386047eda28f952a16768abd2c3decbf88d
Change-Id: I8546189e8475ebff03d0b055ecf7552322b6f51a