[roll] Roll fuchsia [kernel][oom] Trigger OOM response on "terminal" heap alloc failure
A while back, we implemented
https://fuchsia-review.googlesource.com/c/fuchsia/+/645109 to ensure
we trigger a system OOM response (when so configured) when an
allocation failure occurs.  We wired it up to pmm_alloc_page(s).
However, we did not wire it up to pmm_alloc_contiguous, since a
failure there is not necessarily OOM-worthy.  This change updates
cmpct_malloc to trigger a system OOM if it fails to allocate because
it was unable to obtain a large enough block via pmm_alloc_contiguous.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983085
Original-Revision: 155b6170f81bdcca6e9c971f67e67e5639ab9f5b
GitOrigin-RevId: c7fb77930e5a4f20ab26d07c3c6240990e1ebbe9
Change-Id: I194694bdca84c8632c4a13142233fb4112a117a9