[roll] Roll fuchsia [fho] Introduce Connector adapter
A Connector implements TryFromEnv and simply holds on to the FhoEnvironment. It
can then be asked to produce an instance of some other type which implements
TryFromEnv. Basically this is a way for tools to keep the environment around and
instantiate new connections to services etc.

We'll need some other architectural changes in the injector to make this useful
but this should be the final API.

Original-Bug: b/310035296
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981472
Original-Revision: 3dd412d4415f6da0189fed817b1de984b8ad81ca
GitOrigin-RevId: 57ae784300ccc16d4e7bb76ca95a7020cdae85f0
Change-Id: I18a3ad55ee9d36e56d6f0577a4606fea808ffbe3