[roll] Roll fuchsia [brcmfmac] Correct shutdown on SDIO init failure
In the unlikely case that SDIO initialization failed there was a
possibility that the bus object created during initialization would be
kept alive and leaked. For the most part this was not an issue because
the driver would be non functional at this point and unloading it would
free up the resources.

With the introduction of dispatchers instead of threads this meant that
a dispatcher in a work queue was leaked and never shut down properly.
This means that if SDIO initialization failed the next reboot would be
significantly delayed by waiting for this dispatcher to shut down.

Correct this by calling brcmf_sdio_exit if SdioDevice::DeviceInit fails
after creating the bus object.

Original-Bug: b/307600923
Test: Manually induced SDIO initialization failure and observe
successful reboots.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983417
Original-Revision: d01dcbd5b826760349c12bf9be45ec78bcc8d86c
GitOrigin-RevId: 956024f0e89baee9bf20a05d07950f4e67b801ca
Change-Id: Ia31ded3094f7b52ab6224a0c1df1b79bc1d6bd87