[roll] Roll fuchsia [bedrock] Function to open a bedrock capability and report errors
We now have a first instance of an error that is entirely handled in
bedrock code paths, without relying on legacy fallback. That is the Unit
capability/Void source.

This CL builds a little scaffolding so that more bedrock error handling
can be added easily in the future.

It turns out that we're not reporting errors to logs in the bedrock
paths, whereas we do in legacy routing. This CL adds error reporting as
best as possible.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983625
Original-Revision: 8d6a97b4086a521834f4a3a42967411bc84466db
GitOrigin-RevId: b7930656727248ffa504560976beb2f2f01cb54a
Change-Id: If4dbf777fa738938b3eec9464a735a9d54e32caa