[roll] Roll fuchsia [build] Specify single-letter RISC-V extensions separately
This change separates the specification of single-letter
mandatory/optional extensions within a profile from others. We do this
to directly construct the first 'token' of standard extensions in the
-march line, which eases future inclusions of optional, standard
extensions to our ABI (as we intend to do with "V").

Original-Bug: 42075221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981481
Original-Revision: 29806c8b06881b6500bf9561727f79dd5c1a8ef2
GitOrigin-RevId: f5b915fb8e09c29b0dc1ab25377b3a1f996c4dfc
Change-Id: I861aa74303b06918404e1fc4dffdbae98ebf2b6a