[roll] Roll fuchsia [iob] Admit BlobIdAllocator::Allocate() variation to control copying
The initial implementation of `BlobIdAllocator::Allocate()` did a memcpy
of the supplied blob. While this is exactly what we want in userspace,
we need a kernel-friendly variation of this routine that abstracts out
the copy, as the kernel will need to take more care when reading from
the user-memory blob.

Original-Bug: 319501447
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982660
Original-Revision: ae6cec20a62374c499e133315372b1f0a3a128cb
GitOrigin-RevId: 5cce2cd655ea4a4b13cbbea6828d78e976b5d52e
Change-Id: I3ef4bccbc0539219aad8e1b72f1a3d2283aa6405