[roll] Roll fuchsia [fdf] Update dispatcher in FakeGpio
The dispatcher in FakeGpio is set to the one
in libasync.  This change updates it to use
the dispatcher provided by the driver test
framework, if it exists.  Fallback behavior
is to use the libasync dispatcher still.

Testing: This changes test code.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983392
Original-Revision: 89dea85a197db3d594320d48489da212600983b7
GitOrigin-RevId: 91afe390ea4094140606d7b07b5b1138af18b25f
Change-Id: Ic5690b91bd6f72824124923b729973f470d2c5bd