[roll] Roll fuchsia [zxdb][debug-adapter] Fix breakpoint and frame issues
Some debug-adapter cliends needs breakpoints to be verified.
Updated code to set breakpoints as verified when they are
resolved in an attached process.

There was an update to zxdb to send update stack frame notification when
symbolization changes occurred. This broke debug-adapter support as it
result in invalid events related to thread invalidation. Fixed the
invalidation event to be sent only when the frame indexes are not valid.

Test: Manually tested with a debug adapter client.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983419
Original-Revision: 836a350646a79f09c38a91e86b12d25372265238
GitOrigin-RevId: 1ee40c22d053f731d67d0a917c6a956d0245aa67
Change-Id: I4c2160b135169cbc0974875359160211c8711cd9