[roll] Roll fuchsia [assembly] Remove package size check dedup by default
Unless the budget explicitly requires blob dedup, it will be turned off
by default. This is part of the process of making package size checks
product-agnostic.

Previously, when blob was seen in N different budgets, the size of the
blobs would be divided by N to compute a "proportional size". This
proportional size was added to each budget. This caused a lot of budget
churn when packages added/removed/updated blobs, and the share count
changed as a result.

Now, blobs are only counted inside budgets, and not across all budgets,
resulting in N=1 for the most part. This forces the budgets to have an
increased maximum (often 2x), which means they are no longer "relatable"
numbers. The budget maximums are now guardrails to help developers set
size goals/restrictions for themselves, and the product size checker is
a better signal of whether the combination of all the packages of a
product fit within a specific size.

Original-Bug: b/321066699
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/980423
Size-Review: Aidan Wolter <awolter@google.com>
Original-Revision: 694587c4eba6e952dd28a14ba36af8d382781d1a
GitOrigin-RevId: 93cf59e3e6d7e34b1b891225c0559a613efc506a
Change-Id: I956e1d0dd445348b98b9b9d40be01dd04b9f4dc5