[roll] Roll fuchsia [assembly][bt] Add config flag for BT Core packages
Add configuration options for including/excluding Bluetooth core
packages. This is defined as a toplevel enum instead of struct because
there are some packages that are valid with/without BT core and serde
tags provides a concise way to specify this.

This unblocks the effort of moving the #bluetooth-core realm from a
statically defined child of core to a product-variable child.

Original-Bug: 94055
Test: fx build

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950658
Original-Revision: 60b27360be8657eba8eee56a19b162e1bea7d9d1
GitOrigin-RevId: 0e7ac024e6d6e3e20715f7bd0aa650c9b7021a2c
Change-Id: I922b8dd86204a9fb6707b95a06ed0c0b1c1a116e