[roll] Roll fuchsia [wlansoftmac] Add async Tx path tracing
In general, when tracing is off, each tracing call in the code is
"about the cost of a branch instruction" which is on the order of
nanoseconds. When tracing is on, it should cost a hashtable lookup
plus some small memcpy's which is on the order of <200 nanoseconds. In
other words, with tracing on or off, we would not expect tracing to
bottleneck data path performance until we're transmitting at very high
data rates.

Test: Manually recorded a trace
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965580
Original-Revision: 46b33b7e852576926a01599e64cbaabfb968a784
GitOrigin-RevId: 235f2409145cd69e04994640d1cac200a25a77c6
Change-Id: I19ecb515f864516668a0250e72078e424b8f5446