[roll] Roll fuchsia [fxfs] refactor open count flushing logic
Simplify the logic around flushing the file when the open count drops to
zero. We now subtract the open count immediately, and then purge if
required, and then, if we need to flush, only flush once.

The previous iteration of this function had a bug where dirty pages past
the end of the file would cause an infinite loop of tasks to be spawned,
holding active guards, which both actively consumed a core with the
spawning activities, and prevented filesystem shutdown.

It's not clear what race condition in the starnix tests caused this
particular issue to be hit, but it is pretty easy to replicate by
writing directly into the backing vmo of a file beyond it's truncated
size. A test with this behavior is added to guard against regressions.

Original-Fixed: b/318434279
Original-Fixed: b/319348464
Multiply: starnix_host_benchmark_e2e_boot
Test: fx test fxfs-platform-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984112
Original-Revision: 0b113fc581cba813ab4f8be08da9b02b86d01214
GitOrigin-RevId: f6878f3dde5e46427d35a4b6055c0a2219df4d29
Change-Id: I927bd49902f2f9f02b79cc5f31054ac5f7c7ce1e