[roll] Roll fuchsia [component_manager] Update LaunchTask to use Router
Update LaunchTask to use Router that returns an Open.
This removes the last amount of generics from
Sender and Receiver.

The main logical change of this CL is that LaunchTask
no longer needs to be served from an async TaskGroup.
This is because it is now stored in the Dictionary and
called directly. This simplifies some of the code.

Original-Bug: b/310741884
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983356
Original-Revision: f37fb29f8b2f1651e6d81cfbb2e4df2236b2261b
GitOrigin-RevId: 6970eccd0bc3ab50920f23c3b4689e69e6e758c4
Change-Id: I2f1e8d7bafa5fb67e72b731c9a60fa8dbffaa591