[roll] Roll fuchsia [components][dictionaries] Build program_output_dict in sandbox_construction
Move the construction of `program_output_dict` from model/component.rs
to sandbox_construction.rs.

Since we now add the Router directly to program_output_dict, this also
affords for a slight simplification of the trait to retrieve a Router
from a Dict.

Reasons for this change:

- Better organization; unifies more of the sandbox building code
- Prepares for followup change on dictionary extension, which involves a
  more complex procedure for creating the Dict

Somewhat unrelated, generalize the child router to accept an `impl
IterablePath` (new trait) instead of `SeparatedPath`, so we can use that
in the next change.

Original-Bug: 323025721
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/983936
Original-Revision: 5b3ba116e0d333826b41282effdb1446b1abb805
GitOrigin-RevId: 6f5689ff31a17178d435517471ea65dd3dadbd10
Change-Id: Ie9d89d18cb0e2106e43d7f591082d5e841ad5c9b