[roll] Roll fuchsia [debugger] Mode SymbolFactory to the DwarfBinary.
Previously the DwarfBinary (representing the file on disk) and the
SymbolFactory (responsible for decoding stuff in the binary) were
separate. They were stitched together in the ModuleSymbolsImpl
class.

This made it impossible to decode symbols given a DwarfBinary which
is all some classes have like DwarfUnit (this is at a lower layer than
ModuleSymbols). This capability is needed in a followup for Debug
Fission support.

This patch moves ownership of the symbol factory to the DwarfBinary.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/977131
Original-Revision: c99950c4e267e8b5b54e3bb2a265564f003781c7
GitOrigin-RevId: a036960b2b106b793214d37ea6a633952b5eb936
Change-Id: Ib4033799a3342979d174d0c86aae84228ee8edbc