[roll] Roll fuchsia [fsverity] Write merkle attribute in multiple transactions if necessary
It's possible that the merkle tree might be too large to write in a
single attribute. Thus, if the size of the merkle tree surpasses a
certain threshold (defined in paged_object_handle.rs as
FLUSH_BATCH_SIZE), we break the write into multiple transactions. In
order to ensure that we do not waste storage if the power cuts before we
finish writing the entire attribute, this change also adds functionality
to the graveyard that enables tombstoning attributes.

Original-Bug: b/323213914
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976492
Original-Revision: 84b4131069d4d01342ceb10246471e68afafbad6
GitOrigin-RevId: 9963c17990198b58783d3e4d8b5d1b8433128370
Change-Id: I4dbc09acb972795e276956a3c6429a483898fc62