[roll] Roll fuchsia [profile] Use only the role name when searching the profile map
https://fxrev.dev/961878 introduced a bug in the ProfileProvider by
unconditionally taking selectors into account. However, the
MediaProfileProvider always appends selectors to the roles it requests,
and expects ProfileProvider to ignore those selectors when searching the
map of configured roles.

I've verified that this works by booting on a sherlock and observing
that a bunch of roles that were previously marked as not found are now
being set correctly.

Original-Bug: 321037780
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984194
Original-Revision: 065300c32497ca79de49b44f4127b2181722f87e
GitOrigin-RevId: cdb131f78544f0c1df4a49680070ff1cd78e9a50
Change-Id: I344947538e1a4425f24291f65920c97fdb940316