[roll] Roll fuchsia [starnix] Send map information to the bpf verifier
This CL introduce a context for the bpf verifier and fills the context
with the position of all of the maps in the bpf code.
It also change the data sent to the code from the index in the local
object list to the address of the Map object in memory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984432
Original-Revision: 39f9e81a2527deff3a70b55b558385c9532bfeab
GitOrigin-RevId: 661d622bd8c0ad5a520411df1f8e13bb1f973276
Change-Id: Idee4449eb84f0e3c25863e90372db2f0ba7ad5d7