[roll] Roll fuchsia [ffx][usb_fastboot_discovery]: loosen device product requirements
Previously, the check of the USB Interface to detect if it was a valid
Device in Fastboot was very strict, and would require matching on Vendor
and Product for the Interface.

Now, loosen the restriction for device products to be only a warning.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985098
Original-Revision: 6712dce862b74354a45c102de7c5f1f9c9857ac7
GitOrigin-RevId: 2aff52c4c5ea9b4e6692cb3c6dd01f7eca481232
Change-Id: Ic6b9c47b9259b97c61e91505cfa320c504565bb4