[roll] Roll fuchsia [ffx][usb_fastboot_discovery]: add diagnostics
Adds diagnostics context to open_interface_with_serial

Add struct UnversionedFastbootUsbTester, which will check if the Usb
Device with the provided serial number is fastboot over useb. This
differs from the OpenInterfaceFastbootUsbTester in that it does not send
the fastboot GetVar command to it to verify that the interface is up and
"Live" only that the USB Device is present.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984400
Original-Revision: b8ddfde27b894871da30da357b74cc88403f4298
GitOrigin-RevId: e93008950564898c2831564182b552928f9e4753
Change-Id: I1448809495b073cde89894799c88978f41ae408e