[roll] Roll fuchsia [fuchsia-repo] Add support for subpackages
This change updates fuchsia-repo client to support recursively
enumerating the subpackages of a package. This is needed to fix a
problem with pbtool, which was not including the subpackage blobs in the
generated product bundle transfer manifests. This also allows
list_packages() to take in account the subpackages when computing the
total package size.

The `ffx repository package show` command was also updated to optionally
display subpackage information using the new --include-subpackages flag.

Original-Bug: b/319394340
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979877
Original-Revision: 9ee9fc9238e9592a847075088c14047f18ac3945
GitOrigin-RevId: 3b6f7fa2ea37d1a7f3d37af23be88c6a3724eb04
Change-Id: Id8963ca66cc88c4b1fa8aab0e1e9b845eb32a89e