[roll] Roll fuchsia [fxfs] Correct permissions on outgoing dir
We were overly strict on the permissions we give to the initial outgoing
connection to the Fxfs component.  This probably wasn't noticed because
of the custom protocols that we use between fshost and Fxfs.

This fixes tooling such as `ffx component copy`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984208
Original-Revision: 5640b7b20ea3d1cd1c71d1d7def4ec7793b0b5f5
GitOrigin-RevId: af0d85725e42872a506604b895d48b017289d869
Change-Id: Iec983f147c7ff6bd636dd12b3b11305d2ec7e099