[roll] Roll fuchsia [build][bazel] Fix @fuchsia_sdk and @internal_sdk incremental correctness
The content hash computed from SDK export directories was not computed
correctly. In the SDK export directory, meta/manifest.json and the meta.json
files it points to are actually symlinks to files that are located elsewhere
in the Ninja build directory.

Before this CL, update_workspace.py would create a content hash that only
depends on the symlink target values, not the content of their destination.
Hence the resulting hash would not change if the content of said files
changed between incremental builds.

This fixes the issue by using os.path.realpath() to point to the destination
files before returning the list of file paths to be fed to the content
hash function.

Note that this is only correct because there can not be any symlinks in
the final IDK archive (our build system doesn't know how to declare or
package these anyway).

Original-Fixed: 323590606
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985332
Original-Revision: 793da6da4bc0f8002c56a9938f25b259b46cc87f
GitOrigin-RevId: f9cbe06b637a0444cd61faab2eaa356dde808ef4
Change-Id: I2bd602fb8945b5c8da3058d39a6a4c0cd0dd136b