[roll] Roll fuchsia [power_tests] A component to generate arbitrary load patterns
When aligning power and cpu usage, we'd like to be able to generate a
known cpu usage pattern to correlate the power and cpu spike. This
component can be launched with a specified config of intervals to
generate cpu loads for.

Original-Bug: 319699097

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965064
Original-Revision: e9b74bfbe09db8596aa3dc3680f5e082eac8d132
GitOrigin-RevId: 6467f56d7fdde59f1b2b8f868138d789831e18ea
Change-Id: I43d5be66e693fadb5f232c102af1a74ec0643c97