[roll] Roll fuchsia [sdk][vfs] Integrate new C++ VFS backend as shared library
Transition the internal SDK VFS library from a static to a shared
library called `libvfs_internal.so`. Publish as partner SDK in
preparation for use by the existing source set.

This new library uses the in-tree VFS for handling connections, does
not leak FIDL protocol details in the interface, and has better thread
safety overall. This serves a similar purpose to `libsvc.so` which is
used by `//sdk/lib/component`.

Add library preemptively to driver allow-lists, as some drivers still
use `//sdk/lib/sys`. This will cause a transitive dependency on this
library. As part of the development of this library, additional runtime
dependencies were eliminated for both `libvfs_internal.so` and
`libsvc.so`.

Reorganize //sdk/lib/vfs as part of the integration.

Original-Bug: b/293936429
Test: fx test vfs_cpp_new_tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981288
Original-Revision: 61e56ae575e85070940876920b6c54355daf6db5
GitOrigin-RevId: 54578b529dc2aee65d892472f2c3a90eb1519cdd
Change-Id: Id1849cc5d75a11c88544a9dcf8779ea993803274