[roll] Roll fuchsia [fidl] Allow breaking changes to in-dev
Since our partners no longer depend on the in-development API level,
it's not important for us to prevent breaking changes there.

This change was mentioned specifically in RFC-0239 as applying to
`NEXT`, but we may as well change how the checks work on in-dev so folks
don't have to make changes 2 API levels out.

Original-Bug: 323889271
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986153
Original-Revision: 7beba66456c5bcfb46ad3321a6be1419131a56c3
GitOrigin-RevId: ff7fd1c1088a2a1ffea57da3663077c7ce9fbda4
Change-Id: I2bd95540f6170c24c168d14bd1cf19ea4238aa4c