[roll] Roll fuchsia [components] Improve errors when routing from void
When routing from void, a `debug!()` log is logged to the kernel log,
and a "SourceInstanceNotFound" error is logged to the component's log
sink. This doesn't match other kind of availability-related errors:

- Logging to kernel is much less useful than the component's log sink.
- SourceInstanceNotFound is the wrong error

So we change it to a dedicated void source error and add a test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985552
Original-Revision: e00454c11288fd685053c684a840914ba35dd51f
GitOrigin-RevId: b4ff7497be77905132a944056a74593bd987f43b
Change-Id: Ib05748244e6d1d79de172c543cfe3d7418cba40c