[roll] Roll fuchsia [components] Move start/lookup methods from Model to ComponentInstance
This is a prefactor for b/320698181 which needs access to EventRegistry
in the Resolve action.

The idea is to move EventRegistry into ModelContext, but that creates
a circular dependency because EventRegistry depends on Model and uses
these methods. In a following change, EventRegistry will depend only
on the root component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986244
Original-Revision: 6d4a3739277985ea813bb78542baee0463daa2b0
GitOrigin-RevId: 7ce0e5270e2015f3c3138971b6a722ae090e9360
Change-Id: I0bef7961f1a2b175fba8f293f4a03c0f1b7949cf