[roll] Roll fuchsia [bazel] use target_name for build-id dirs
The bazel SDK was writing the build_id directories to
package_name/.build_id but this was causing collisions when there were
multiple packages defined in the same BUILD.bazel file because they
would both try to write to this directory.

This change puts the build id directory in ctx.label.name
+ _build_id_dir which makes sure it is unique to the bazel package and
target. Additionally, we had to update the bazel_action to look for the
build id directories at the new location. The code was updated to use
scopes instead of strings so we could piggyback off of the name
templating that is already built into bazel_action.

BUG: b/322752815
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986110
Original-Revision: 7e6e8084a0233740019020d7da250c20f19dd6a4
GitOrigin-RevId: 242c06177084390df85773b4e490e885c41dcef4
Change-Id: Ib0d763cbe98ed79a205bfc37e74107c4a7bfef7f