[roll] Roll fuchsia [debugger] Log output should respect OutputEnabled
When getting log notification requests, we should respect the
OutputEnabled bit that is checked in the `Output` path, but not the
`Write` path. When operating in embedded mode, we really don't want
spurious DebugAgent logs getting mixed in with the logs we're likely
streaming via fd_streamers.

This was particularly noticeable if you ctrl+c'd rust tests being run
under fx test with any of the debug options and then getting a bunch of
"Process with KOID not found" messages as DebugAgent was cleaning up.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984586
Original-Revision: 018a7e0b7837e4f7cdfd4404ca9d24e8bbf0166d
GitOrigin-RevId: dbf1a3c26f3fcceb929f50ff7b02604c4bfcd175
Change-Id: I008e72cc5554aa2d62ee7a09ed71c76c8157eac0