[roll] Roll fuchsia [kernel][phys] Introduce RISC-V feature set abstraction
This change introduces arch::RiscvFeatures, a container abstraction
around the set of RISC-V features of interest to the kernel. This is
populated in phys as the lowest common denominator of features across
all hart ISA strings and is included in the hand-off.

As a first use, the kernel's RISC-V feature detection constants -
previously hardcoded to true in wait of a programmatic means to query
this info - are updated to be defined in terms of the new handed-off
feature abstraction.

This will be soon be used to conditionalize support for select RISC-V
features in the kernel.

Original-Bug: 124336
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955485
Original-Revision: 6911a79b73086ac2b281332c32c7cd0bab7ed891
GitOrigin-RevId: 0f6c82bff479c6b7880f369f6f8360f4770fe3e0
Change-Id: I129011672b1fd87dbe366c1e1faa298de0e39b11