[roll] Roll fuchsia [caching-package-server] Support Inspect
Adds a method to register a lazy inspect node that will expose open
packages and their connection counts.

Also adds an option to provide the serve method with a RootDir to
speed up instances in which the client already has one and the package
is not open.

Original-Bug: b/307977824
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979317
Original-Revision: 9876c346765a999b6b2126ce50083a2b3b7749ba
GitOrigin-RevId: ade4237a591c0aecb18293d0fed3c2ebe9f4a96c
Change-Id: I52e58b14df51ca903d4ea005c6ad7e06566c379a