[roll] Roll fuchsia [ffx][fastboot]: add rediscover function
InterfaceFactoryBase now has function rediscover(&mut self) which is
intended to be used during the flashing process in order to rediscover
the target after it has been instructed to reboot to the bootloader.

This code is not wired up to be used yet.

The current implementation of the reconnect functions assume:
  1. A target will never change fastboot connection mechanisms. For
     example, it is assumed that if a target is connected via USB, after
     the bootloader reboots it will continue to be available via USB
  2. A target's name (for fastboot over TCP/UDP) is consistent across
     reboots.
  3. A target's serial number (for fastboot over USB) is consistent
     across reboots.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984588
Original-Revision: bb4021efbcafdb23b5817a65cbd32f5207a6ee30
GitOrigin-RevId: 8475fc7b8a1549b32a9f2b337878a0b920b7b25b
Change-Id: Ib6e40e028e036d61d729556669cbcca040200d44