[roll] Roll fuchsia [starnix][fuse] Truncate writes to transfer size
The `fuse_out_header` has a `len` field that describes the length of the
payload data. If that length exceeds the total amount of data written in
this call to `write`, we should truncate the size of the payload to the
available memory rather than return an error.

Original-Bug: b/323001474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986474
Original-Revision: c1f3694244b7db7c31423a24281dae4690c2ac92
GitOrigin-RevId: 955193814cc982fc76c53114058c656067574f89
Change-Id: Ib28c4a2ce2412960e9106a8350e2113bd61ee7f4