[roll] Roll fuchsia [starnix][splice] Do not eagerly lock file offsets
Our implementation of `splice` was eagerly locking the offsets for its
operands. This locking is usually not necessary. For example, one of the
operands is always a pipe, which does not need to lock its file offset.
Similarly, if when the operand is a fuse device, the lock isn't needed
because those devices do not use offsets.

This CL resolves a deadlock which can occur when multiple threads are
reading and writing a fuse device using splice. Now, we take the offset
locks using the normal offset locking machinery, which is aware of when
various operations block.

Additionally, this CL removes the `read_raw` and `write_raw` operations
on files because these operations no longer have any clients.

Original-Bug: b/323001474
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986421
Original-Revision: 72ce1306030c62011b116688fde1c033324ce42b
GitOrigin-RevId: 6f2587a9ee801fb827eb6a80f7861d7523e62d43
Change-Id: Iee437fbf9e84758e2738996cb5ff617b67ddebfa