[roll] Roll fuchsia [ffx][flash]: wait in a loop for Target to Reboot
When first executing the flash command, before the flashing code is
executed, we need to ensure the Target is rebooted to the bootloader.
Instead of waiting for a flat 10 seconds (brittle), wait in a loop.

If the target is not discovered in Bootloader mode in 20 seconds, warn the
user.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986098
Original-Revision: 61712c1185bd5c68df3f775167185cd4c69bc543
GitOrigin-RevId: 1f2a642fe9f9e0e5f61371246ad395e2cb1dba16
Change-Id: I3d9ccd0226ea9134e8a55acacbe17ffc6b62e944