[roll] Roll fuchsia [netstack3] PacketSocket TODO for Pure IP devices
Remove an incorrect comment that stated Pure IP devices did not need to
deliver packets to packet sockets. That functionality is supported on
Linux and gVisor, and will be required for Netstack3. Replace the
comment with a TODO.

Original-Bug: 42051633
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987074
Original-Revision: 36663cb3b9f6f5bdbd20c5bff18d721edc6960e2
GitOrigin-RevId: b3cc942b0b691c72ec84b2ea31b1a66e8ba28ed3
Change-Id: I490514763d21813b03dc1bbc75577442c3a86bee