[roll] Roll fuchsia [ffx][bootloader]: wait in a loop for Target to Reboot
When first executing the bootloader command, before any fastboot code is
executed, we need to ensure the Target is rebooted to the bootloader.
Instead of waiting for a flat 10 seconds (brittle), wait in a loop.

If the target is not discovered in Bootloader mode in 20 seconds, warn the
user.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986232
Original-Revision: 5ee44e31b6a2cfb830fe5cc3697ea625960b2293
GitOrigin-RevId: 5bf1ae450ab9c0e82775357379e62c524eabd069
Change-Id: Ib61dc28f635c37888f4a61518a35b93d280e20e3