[roll] Roll fuchsia [netstack3] Add optional local delivery device
...to ResolvedRoute which carries the device that the destination
address is assigned to in the local delivery case. This enables
`new_ip_socket` to bind the socket to this device if either the source
or destination addresses require a zone ID. Prior to this change,
the socket was being bound to loopback, which is incorrect.

Return an error from resolve_route if the source and destination
addresses are assigned to different interfaces, and one of them
requires a zone ID. Note that Linux does allow this, but exhibits
strange behavior. This case is difficult to support correctly, so
we prohibit it for now.

Skip the integration test that pings an IPv6 LL address assigned to
oneself, which used to fail due to a routing error, but now stalls
forever because delivery of the echo reply cannot make it back to
the ICMP socket due to fxbug.dev/42084713.

Original-Fixed: 323958072
Original-Bug: 42084713
Test: fx test --test-filter '*lookup_route*' netstack3-core-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982054
Original-Revision: 9e64948af27f4e7e75ce88b7572f2796f2dfe0ec
GitOrigin-RevId: 7880a2172cbe6fa064b84aafa05ced8ce842e130
Change-Id: Iaa5003861023f2ac8eaf717e773fbdbd2ec9b694