[roll] Roll fuchsia [build] Disable -Wextra-qualification
This works around an issue with -Wextra-qualification, where fidl
generated code is causing this warning to fire and break the build with
newer versions of clang. It is currently unclear what the underlying
issue is, but this flag can be removed once a long term fix is available
in the compiler or in the Fuchsia tree.

Note: -Wextra-qualification warns about unneeded qualifiers on members.

Original-Bug: 324114504
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987955
Original-Revision: 389ebb6eff0b2f27b5fe9fa58c04031f25103ab2
GitOrigin-RevId: 1e64127ea5292ec4b97299b9cdaef12c2c3aa489
Change-Id: I77d6e84b780c6ae461113cc2f597b720935d2a1b