[roll] Roll fuchsia [role_manager] Avoid unnecessary get_info call
We use zx_object_get_info calls to get information about the thread/vmar
we're setting a role on, but we only use this info to emit log lines to
the syslog at the DEBUG log level. This means that the results of the
syscall are functionally unused if the log level is higher than DEBUG.
This CL fixes the issue by making the syscall invocation conditional on
the log level.

Original-Bug: 323262398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987900
Original-Revision: 651117600670efeb88a97277927357f2d3e6084c
GitOrigin-RevId: e6f75f9766bcde29b7313b2c6d2e157e24e0604d
Change-Id: I0404053e0c2fbe57de33a6f5551d6e26634c4248