[roll] Roll fuchsia [publish] Fix `fx publish`.
`fx publish cache` is broken because it tries to run
`fx build build/images/updates:prepare_publish`, but the GN build graph
is unaware of that target by default.
So, add it to `//:root_targets` for now.

This is intended as a quick fix, with a follow-up cleanup cl if
feasible.

Original-Bug: 324385881
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988152
Original-Revision: 5ec569507be5f66fc6c2651c4d21d955c6bfed50
GitOrigin-RevId: ad5422ee809140e0623ad71b5c0a3d0fef0d844b
Change-Id: I4707d288694f405fc90fd1ec19165cb384af9174