[roll] Roll fuchsia [fx][serve] Warn about running foreground server not being detected
When the ffx foreground repo server (currently in beta) is enabled (by
setting the environment `FUCHSIA_DISABLED_foreground_repo_server=0`),
it is not currently not possible for `fx serve` to detect if an ffx
foreground repo server instance is running, or if something else is
merely using the same port. Warn the user about this, and suggest
the workaround to disable the ffx repo server listening on a port.

Original-Fixed: b/323583171
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988193
Original-Revision: 9d87b811b86c351b236adbe7e09509efd6240420
GitOrigin-RevId: 973b63fd5f61a7d528a2ca6673c0e1553cc57c42
Change-Id: Ia7a8c430d617d55bb4d178cd03a8d88e9cb828a2