[roll] Roll fuchsia [netstack3] Convert filter state to core state
When fuchsia.net.filter/NamespaceController.Commit is called, after
applying all pending changes to its own copy of filtering state,
Bindings then attempts to convert its state to netstack3_core::filter
state types, which are generic over IP version and leverage more type
safety in their representation. This involves validating that:

 - There is no path from a routine installed in an IPv4-only namespace
   to an IPv6-specific rule, or vice versa.
 - There is no path from an IP routine to a NAT routine, or vice versa.
 - There are no cycles in any of the routines.

This change also moves some of the CommitError variants that are
reported as specific to a particular Change to instead report the
resource that caused the error, because those error conditions are
holistic properties of the entire filtering state (dependent on which
routines jump to which other routines) and can only be validated against
a set of changes.

Original-Bug: 318738286
Test: netstack-filter-fidl-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982554
Original-Revision: 2f4365cb23774ec9da19fa126bf448bc04b9e822
GitOrigin-RevId: c8892882f154052b3bd66697b88a7a3dbe395303
Change-Id: I0749ea219491a5be5bfca351285c72ac4128d21b