[roll] Roll fuchsia [netdevice] Support binding to FIDL parents
The migration to FIDL added most of the infrastructure to bind to FIDL
parents but without any actual FIDL parents to test against, a couple of
things were left out.

Add the missing bind rule for FIDL parents and indicate that the network
device driver uses the network driver service. Without the use statement
the binding would begin but could not complete because the right
capability could not be found for the service.

Test: Verified network-device binds to both Banjo and FIDL parents.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987220
Original-Revision: 42bc7d5183463039e57fb1570f0bad06dbdb94c8
GitOrigin-RevId: 37ff971c6789be2d10a8a9067446ded280724bd6
Change-Id: I60a90b65dd5a2decc07086dd53ea875f91ef7c57