[roll] Roll fuchsia [Scudo] Add custom scudo config
Now Fuchsia uses a custom config defined in Fuchsia repo instead of
relying on the config defined upstream. The new config is the same as
the one used previously, except that `RegionSizeLog` was reduced to 27
bit on RISC-V. This is a workaround to make Scudo cope with the reduced
address space on RISC-V (particularly in restricted-mode, e.g. in
Starnix).

Original-Bug: b/321064019
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986233
Original-Revision: 3156c029aea8d0b730b8eda660a4ba3f5f5c1fe0
GitOrigin-RevId: 185b0d29fb61f13c73ae2c1b50e51823523fe134
Change-Id: Idcf9f41b05c2c46781672fb93fd1d9ac4ea4e758