[roll] Roll fuchsia [fxtest][debug] Allow status when in debug mode
zxdb will now not mix the outputs from debugging actions and fx test.
The result of this is the user will see the fx test status output as
normal until the debugger pops up because of some test failure. When
this happens the status output will be hidden and not mix with the
debugger's output. When finished debugging the test failure, the status
output will be seen again for the rest of the test suite (or until
another failure brings zxdb back to the foreground).

A short demonstration can be seen at
asciinema.googleplex.com/5652956951085056

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987972
Original-Revision: 65301d641e3201a27c75165cb2e10a4a8d6ca7ef
GitOrigin-RevId: 07c751826afddf227f142a47d8d899b6e1979058
Change-Id: I2104909887ccbca9bebf13b9f3ed724caba1c611