[roll] Roll fuchsia [bazel] add fx bazel-refresh-compdb command
Adds an fx command which updates the compile_commands.json file with all
of the deps of a given bazel label. This command is a temporary solution
which gives us linting/error reporting for vscode.

The tool can only b called after the bazel sdk is generated or it will fail.

The code will attempt to convert what bazel thinks the paths should be
to something that maps to the fuchsia tree.

In the future this will be built into fx set but we cannot do that until
we can call bazel commands without first generating the bazel sdk.

To run the command:
fx refresh-bazel-compdb --label //build/bazel/examples/hello_cpp:pkg

BUG: b/323245932

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987966
Original-Revision: 16469cba3903e90b54265f4cf0232f176cc4d2ff
GitOrigin-RevId: db8b122f47f769124917d2d9cf8ea880c8b5b05e
Change-Id: I9f25719b62769e98c4200c2c2e2f1424469f6d6a