[roll] Roll fuchsia [scheduler] Copy role library to SDK directory
Copies the scheduler role library to the SDK directory. This library will
be used by clients to set roles on threads and vmars, and abstracts away
some of the boilerplate code associated with making FIDL requests to the
fuchsia_scheduler APIs (this is currently the `ProfileProvider` service,
but will be updated to use the `RoleManager` service once all of the
callers have been migrated to use this library).

This change also migrates all in-tree users of the scheduler library to
use the new SDK based library.

This CL does not delete the role library from `//src/lib/scheduler` as
there are existing OOT users that must be migrated first.

Note that this change keeps the scheduler library internal only. Future
CLs may change that after an API review.

Original-Bug: 323262398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988552
Original-Revision: d637227a959daa77dd8490b96e79c1efeda5aa1d
GitOrigin-RevId: 6f3ad69de640d4cb04bb9c612c5b15d2f5605944
Change-Id: I5660904c659e8b782a1b77c50dd3d76e483edeb9