[roll] Roll fuchsia [fdf] Add transport to offer table and set prop
Adds a new offer type in the fuchsia_driver_framework's NodeAddArgs fidl
table. This has the ability to convey the transport type of the offer.
Then the driver manager can take this information and add the property
for the offer and its transport automatically.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/979191
Original-Revision: 6d55f08d5aef0fc5612c810041ad215aa3699447
GitOrigin-RevId: 218a2d96fd603a583d7cc18881ec882def3d231a
Change-Id: Ifc82e22bd0926df286f40a9059a4933aa105e472