[roll] Roll fuchsia [virtio-guest] Extract VirtioPciDevice class in v1 driver.
The VirtioPciDevice class encapsulates virtio device implementation
details that are not specific to the virtio-gpu device type.

This CL is focused on moving the code out of the Gpu class. Some
details specific to virtio-gpu, such as the virtqueue setup, are moved
to the VirtioPciDevice class. Follow-up work will revise the API
boundary.

Original-Bug: b/42182631
Test: Booted core.qemu-x64 using ffx emu start --engine qemu
Test: ffx target ssh display-tool squares
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988037
Original-Revision: 4855f9ae5fd8591766773ec50dcfa445217eaee1
GitOrigin-RevId: 2caf351d18db3318f81720205c9d02b462fcf16e
Change-Id: Iafcb5db44b3bbc24d08ae65f7e75b782fcea5f27