[roll] Roll fuchsia [bedrock] Add a Capability enum
Use a Capability enum everywhere instead of AnyCapability.
This gives type safety when converting back from the types
within a Dictionary and from within the FIDL registry.

AnyCapability still exists because Router is defined
outside of the bedrock library. Once it is moved into the
bedrock library the type can be deleted completely.

The Opaque and Lazy capability types are removed because
they are unused and they require templating so it is harder
to place them in Capability.

Original-Bug: b/324606473
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987774
Original-Revision: c039a6744395ae2c98381f5bd1493d30fd62d360
GitOrigin-RevId: 3276075088686420cc9ee812210be7cc850571fd
Change-Id: Ia69f4119054c59aa650bdd643165ea52e7070f5e