[roll] Roll fuchsia [components] Remove capabilities from registry when task finishes
When a capability is converted to FIDL, the sandbox library inserts
the Rust Capability object into the registry along with a Task that
typically serves a FIDL protocol.

This leaves an extra capability in the registry when the task ends.
Since capabilities are shared references, this means capabilities
are never fully destroyed after being converted to FIDL.

FIDL conversion now removes the capability from the registry when
the server task finishes.

Original-Bug: b/323582737
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989358
Original-Revision: 36faca9f3affc23981117da1d76d3c24cf849747
GitOrigin-RevId: 41875b3d1a6040be8e3a3c352106626baaff065f
Change-Id: I1255ba0040def8d2c745f35679da25c08fa38ea1