[roll] Roll fuchsia [debugger] Output a helpful message when leaving embedded mode
It can be jarring for new users to be dropped into the debugger with an
exception location that might not be in their own code. Print a bright
warning message saying what happened and a recommendation for what to do
next.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988035
Original-Revision: 2cf7571bee5634dc0504e7f85d928ac89350388a
GitOrigin-RevId: e5d5fda3dcaca6e34701dd7ec3453d9cac491475
Change-Id: If673ef2bf85b5b73c4b7618baf23b0e2b7c09c64