[roll] Roll fuchsia [kernel][vm] Fix ZeroPages for SnapshotModified
Fix the bug in Snapshot-Modified in which ZeroPagesLocked applies an
inappropriate optimisation. Previously, ZeroPagesLocked would check if a
clone's parent was hidden, and if it was, a gap was inserted in lieu of
a marker or zero page.

This doesn't work for Snapshot-Modified clones as a hidden cow pages can
now have a pager-backed parent. This can result in pages from the
pager-backed root being leaked on reads to a range that has been zeroed
with a gap.

Original-Bug: 323447923
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984675
Original-Revision: 7d0d0bd75ff7266487cd5e6afe2349aaf1dace10
GitOrigin-RevId: f34a7131c324eefc6ae4448b6d89ce56c1a2f86b
Change-Id: Iad138669a85f3451d7a085777b13013fdd6e36f8