[roll] Roll fuchsia [driver][sdk] Use SDK to build Example Board Driver
Fix the dependency graph of the devicetree libraries. Now the example board driver can compile, but the test for it is not wired up yet. This will happen in next CL.

Also included in the change is the support of cc_source_library depends
on a bind library in SDK.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988378
Original-Revision: 7cc55aa075446c7355f9fe76f35369bfcab595a8
GitOrigin-RevId: 8a19d63e697c8e98adba05561a213ec78975ed0f
Change-Id: Iba4eb6e2028515f919f84de174a6b4b6ac2a28f2