[roll] Roll fuchsia [netstack3] Allow configuring DAD transmits
Introduce a configuration to fuchsia.net.interfaces.admin allowing the
number of DAD transmits to be changed, this can help deflake some tests.

We don't implement that in Netstack2 because a) we can just disable
flaking tests in NS2 if they become too much of a nuisance, b) we're
experiencing most flakes on RISC-V builders, which run Netstack3, c)
gVisor is not set up to configure DAD transmits per-interface.

In Netstack3, we have an impedance mismatch between the FIDL API
organization and the core API because `dad_transmits` already exists
there and is inside the IPv6 device configuration lock. The natural
follow from the FIDL form would be to put DadConfiguration inside core's
NdpConfiguration, but that is held by the device layer and far away. We
can revisit this if we add more configuration and a clear and obvious
improvement becomes available. Part of desire to split that is already
captured in https://fxbug.dev/42173351.

Original-Fixed: 323453776
Original-Bug: 42173351, 42182861, 42077260
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990262
Original-Revision: e27560e6f70f359d1483fb453a1d00271993bc96
GitOrigin-RevId: a6f9f66748428926165031628a66f53ebde1adf7
Change-Id: Ia3eccb31672f6c0b67d24bc49c38e5b69700a28a