[roll] Roll fuchsia [scenic] Removes a dep from bin to component
I am not sure why a binary depends on a component. Looks wrong.
If we want to ensure that a component is present, we need to ensure that
elsewhere, e.g. in the scenic component dep.

Original-Bug: b/324512578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990939
Original-Revision: 86e5795b5beb44e50b600aa4c4132e45fb9e951e
GitOrigin-RevId: 01230193491118a7a884f62b19cd387db154ffc5
Change-Id: I8bee5752ead5ae62ee3ed3e6a1f01691078b1135