[roll] Roll fuchsia [virtualization] Switch to setting profile by role
The virtio_sound driver was explicitly constructing and setting a
deadline profile using the deprecated `GetDeadlineProfile` method of
`ProfileProvider`. This change does a couple things to remedy this:
1. Creates an explicit role for this deadline profile.
2. Modifies the `virtio_sound` driver to set this profile using the
   new SDK scheduler library that calls `SetProfileByRole`
Unfortunately, the virtio_sound_driver will still need to know the
deadline profile it sets because it uses it to compute some other
things. This can be replace with output parameters once we switch over
to RoleManager, but I'd like to keep this change as mechanical as
possible, so that's left for a future CL with a TODO in this one.

Original-Bug: 323262398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990516
Original-Revision: a2e091b92871e657cfb066852be56e53e7bc1110
GitOrigin-RevId: 4353d4eead4202d894e4a066bd5325c17d4cef3a
Change-Id: I899fb80b684ab4423705860ba948b214ee8a506e