[roll] Roll fuchsia [ffx][fastboot]: add FastbootUsbLiveTester trait
There are two distinct steps when evaluating if a USB device is
"discovered".

1. Does the USB interface match what we consider to be a valid fastboot
   interface? (i.e. Says it Speaks the Fastboot Protocol, and is a valid
   Product/Vendor)
2. Is that USB interface ready to respond to Fastboot commands?

Now for the FastbootUsbWatcher, add a "tester" parameter
(impl FastbootUsbLiveTester) which is used to test if a USB Interface is
"live" aka ready to respond to Fastboot Commands.

The new flow of the usb discovery loop is

1. Find USB Serial Numbers
2. For Each Serial Number we have not found yet
  a. Check if the USB Interface for that serial number is a Valid Fastboot
     Interface (if not, skip it)
  b. Check if the Device on that USB Interface is "live" (if not, skip
     it)
  c. Send a Discovered Event for that Serial
3. For each Serial Number not found from our previous loop
  a. SEnd a Lost Event for that Serial

The previous flow would conflate the "Valid Fastboot Interface" step
with th "Is Live" step, erroniously.

This new approach gives finer grained control over the discovery loop.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991368
Original-Revision: 3ec507274d4ac864ad31b9431924e556656a2d61
GitOrigin-RevId: 252920552dcf82cbd65fc51936d3a09f8ae63b3e
Change-Id: I425eccff9c73068c459fc15f73414f9d569c659a