[roll] Roll fuchsia [wlan][policy] Retain recovery events for one day.
This change causes defects that are observed and recovery
actions that are taken to be preserved for 24 hours.
Previously, these events were immediately dropped as their
historical records were not actually used in making any
decisions.  Recovery is still logging-only.  The effect of
this change is to ensure that wlancfg logs when recovery
*would* occur if it were enabled.  Some small amount of
memory will be consumed in storing this extra information,
but it is not expected to be overly burdensome as the
events that are tracked are extremely infrequent.

Test: fx test wlancfg-tests
Test: fx test hw-sim-tests
Test: IOU - fxr/991646
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991645
Original-Revision: d5ba64fa16c59e8549a087f941f9801a8a7b473a
GitOrigin-RevId: 56d3a381cf776744ee81bf88266778a686df98f4
Change-Id: Iadea8f23a643038dd8fb25512d2893349d10b958