[roll] Roll fuchsia [starnix] Fix wait notifications using ptrace
Previously, if a task was being traced, a notification would only be
delivered to its tracer.  This CL adds a data structure that tracks the
original parent of the traced task, and notifies it after the tracer
has consumed the notification.

Original-Bug: b/299348370

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986097
Original-Revision: c8169764da7459fe4a96947bc398af05851d20d1
GitOrigin-RevId: 7294c0dac0853fd8659b90059d14cc3a73ce35ee
Change-Id: I3925569529e905369ea70ba124b07fef5a8f115c