[roll] Roll fuchsia [assembly] Size aggregator always emits all outputs
We want to be able to collect and upload size results on builds that
only run a product size checker. This means that the "budgets" and
"verbose_output" would be empty. This CL updates the aggregator rule to
generate empty files when the inputs do not contain these files.

Original-Bug: b/319519133
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991363
Original-Revision: 3b0a948debbcdbb8643fd686fecb7b5c959c5d52
GitOrigin-RevId: 8772bd721170dc0ac9ecb04a9073ac598d218a51
Change-Id: Ief0ee9867b0db40ef6c11d4cca4309914f9e4dc9