[roll] Roll fuchsia [flatland][test] Do not perform blocking wait on FIDL messages for MockDisplayCoordinator.
CreateServerWaitingForMessages() creates a thread that actively
listens to the FIDL channel. It effectively bypasses the async
dispatcher and dispatches all the FIDL calls to the wait thread.

This is fragile and it heavily depends on the implementation
details of the fidl::Binding, and it breaks the threading model
of the FIDL server. This is not compatible with the new C++ FIDL
bindings.

This change:

- Removes CreateServerWaitingForMessages() and creates a separate
  async Loop (running on a separate thread) to dispatch all the
  FIDL DisplayCoordinator calls.

- Replaces the normal Nice mock with a Strict mock (see [1]) so that
  all "uninteresting calls" which we do not specify EXPECT_CALL
  will trigger a test error.

- Explicitly specifies the Times() of each display compositor
  call in EXPECT_CALL().

- Fixes some other coding style (unused argument, incorrect include)
  issues.

[1] https://google.github.io/googletest/gmock_cook_book.html#NiceStrictNaggy

Test: flatland_engine_unittests
Original-Bug: 42150597
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989359
Original-Revision: d595bad83f7ac06b3d7e8d8c65cd521fa6fb4237
GitOrigin-RevId: 908f69f293f778cf87410c81686787cdfaed92da
Change-Id: I552ffc3c583aad43720c1588bb678bf8da2f22c6