[roll] Roll fuchsia [realm_builder_server] Fix ordering/timing issue
Before this CL, the tests would drop the runner server end.  A watcher
would look for peer closed on the client end and clean up which, if it
were to happen in time, would cause tests to fail.  I had a change to
fuchsia-async which caused this to trip.

This changes it so that we don't close the server end which stops the
tests from failing in that way.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991752
Original-Revision: adc73daf144791fb6f90ac8e8bcdd92fed8fbcb5
GitOrigin-RevId: 08196a067f8373531bb72189d72aa7659080cbf6
Change-Id: I994a97d64f64cdebc75d2301e723148abbaf811d