[roll] Roll fuchsia [memory_monitor] Find Starnix kernel cutoff at runtime
The cutoff address between the restricted space and the Starnix kernel
depends on the architecture. This change relies on the fact that shared
processes, as used by Starnix, have two root VMARs, one for the
restricted space and one for the Starnix kernel, to determine this
cutoff at runtime, instead of having it hardcoded.

FIXED=319838618

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981333
Original-Revision: ee1346682ba670899592fbe192fa75b2af271b93
GitOrigin-RevId: 2637aca5e9892e0f368f022c52b82cc95528135c
Change-Id: I98001bd099439b020997e26be270698d96586c3d