[roll] Roll fuchsia [check-licenses] Skip zircon/prebuilt
It was deprecated in fxr/735867 by adding a .gitignore entry, so
check-licenses can safely ignore it.

That CL mentions that the contents of that directory may still exist
on developer workstations, so a skip entry is needed in order to
ensure folks can run the tool locally.

Original-Bug: b/325298549
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972996
Original-Revision: 3bdb0123f75129193a6509c9ebec692ab600d0fd
GitOrigin-RevId: fca1c64985ee70544530ce71fe7faf73c1b0fbcc
Change-Id: I8befbae58cc62f1a8a931f16a251a991c8a8c109