[roll] Roll fuchsia [build][riscv64] Introduce `riscv64_enable_vector`
This change introduces a GN arg to gate whether we enable RISC-V vector
instructions. We have a use-case in which we'd want to disable this and
are also currently investigating a bug in the case in which we intend to
have it enabled; this can be set to `false` locally in both of those
cases.

Original-Bug: 42075221
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987904
Original-Revision: e35edecf5750fd65339333d7967b8c3ab634bf94
GitOrigin-RevId: afa9806eb661abac5f73acc39cc16c7ffa405176
Change-Id: If0c23749eb635c67effc7b50c0c8aab5a74ae390