[roll] Roll fuchsia [bt][snoop] Open bt-hci as Vendor
We moved from publishing devices as Hci FIDL and instead
publishing as a Vendor to enable using Vendor FIDL in bt-host
component.

We switched this recently, and now bt-hci-broadcom panics if
it gets an unrecognized message.

Open the device as a Vendor FIDL instead and get the Hci FIDL
from it.

Original-Bug: 325329669
Original-Bug: 325302214
Test: build nelson and use A2DP sink
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992712
Original-Revision: f24929d27a6435577390568a6327e04f64c3cb29
GitOrigin-RevId: 15d2e8ea188c62c8025499c0fa14a29266cbf995
Change-Id: I9c093224bf78a2a7d337a9cd9590386c1e890ca1