[roll] Roll fuchsia [netstack3] Validate and install filters in Core
Add a witness type to netstack3_filter, ValidState, which witnesses to
the invariant that the contained State has been validated to uphold
properties that are not guaranteed by the type definition: namely, that
no rule has a matcher or action that is invalid for the context in which
it will be evaluated.

For now, the only possible case of this is a rule matching on the output
interface in the INGRESS or LOCAL_INGRESS hooks, or a rule matching on
the input interface in the EGRESS or LOCAL_EGRESS hooks. Once we add
actions like TPROXY and NAT, those will also have to be validated.

To allow Core to provide a specific error to Bindings on validation
failure, add a ValidationInfo type parameter to netstack3_filter::State
that specifies ancillary state to be stored for each Rule. This type is
zero-sized for validated state at rest in Core, but Bindings provides a
RuleId with each rule when installing the state so that Core can report
which rule was invalid when validation fails, and Bindings can in turn
report that to the caller.

Original-Fixed: 318738286
Test: netstack-filter-fidl-test
      netstack3-filter-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986254
Original-Revision: d20db4539cbed18395a59d807e98c30f82a8f03e
GitOrigin-RevId: e816086aba8105b76e23a62808016001ebc24bbf
Change-Id: I6a649c80206abbffbc7916a44ebd998513284f98