[roll] Roll fuchsia [ffx] Add In-use handle value type to Playground
The Rust bindings for Zircon handles, and the value type we get from
FIDL codec, both treat handles as single-owner. In the shell environment
this doesn't quite work. To get around this, we add an in-use handle
value type, which is sort of a ref-counted container with a handle
inside of it. We can duplicate this value as much as we want, all the
copies reference the same handle, and any one of them can steal the
handle if we need to send it through a channel for example, with the
others inducing graceful failures if they're operated on further.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987295
Original-Revision: d8d7852ca63ed21afa2d6ddbd7ac2434038d02f1
GitOrigin-RevId: 0e5617aa2407377708dfb31e24220d6d8d063668
Change-Id: I74ca102e99490e6c83dc47c55c1249c8b5586ee3