[roll] Roll fuchsia [performance][docs] Update docs for creating metric name expectation files
Update these docs to work for Python-based tests.

Drop the description of doing this for Dart-based tests.  We generally
won't be adding new Dart-based tests, so we don't need to complicate
the docs with covering that.

These docs are generally complicated by the fact that GN will complain
if you specify (in GN) an expectation file that doesn't exist yet.  An
alternative here might be to add a GN argument for telling the tests
to create these expectation files.

Original-Bug: b/307980390
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992753
Original-Revision: a936253290fe334609303572cdc1d3e385d5ce92
GitOrigin-RevId: a024206a3952f2f7eedb00039c654b377cb09d58
Change-Id: I1ba3ce6b90b064c4271ec0df069400bc9ac6aae1