[roll] Roll fuchsia [kernel][arm64] Use <lib/arch/arm64/exception-asm.h> fully
This finishes the transition to the assembly macro scheme with
.vbar_table and .vbar_function already used in the phys code.
The instruction sequence and layout of the vector code is not
changed, only the assembly source uses less boilerplate and some
symbol and section names are changed in a uniform scheme.  This
paves the way for moving the code around and further macroifying
things more easily, including the possibility of multiple
variants of the vector table with minimal source duplication.

The only actual change in the disassembly is the "vector number"
printed on the panic path for invalid exceptions, which simply
uses the prescribed offset from the vector base address to
indicate which hardware vector it is, rather than the previous
idiosyncratic set of numbers.

Original-Bug: 323463901
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985732
Original-Revision: 3f10ea9aca5fb8495ea5827e0c1f48fb29e7a82d
GitOrigin-RevId: f13402a1830093e6d1a5e93fb729db01106cd14d
Change-Id: I0265569605a8b10f7754cb124938560ba125d861