[roll] Roll fuchsia [idk] Simplify archive step.
While debugging fxbug.dev/313682083, I was surprised to find that the
content of the IDK archive differs quite significantly from the content
of the directory it's _supposedly_ an archive of. This is because we
archived the directory by invoking `merge.py`, effectively telling it to
"merge the completed IDK directory with itself, and then zip it up."
This lead to a bunch of confusing side effects (like some parts of the
IDK metadata being included twice), and generally seems to be the wrong
way to go about things.

This CL takes a much simpler approach by walking the contents of the IDK
directory, resolving any symlinks, and handing that list of files to
tarmaker. This _does_ leave us vulnerable to bugs involving stale files
left over between builds, so I think we can do better than this, but
not without substantial changes to merge.py, which I'm not ready to
undertake at the moment.

Original-Bug: 313682083

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987919
Original-Revision: 48c26e9278bfd317d2c4cb15d35183dd07609168
GitOrigin-RevId: a91cbaa40068d3cdadb48d3c27a32c2f5390e7a4
Change-Id: I8197498c1b1ee96a30d224795c3cabc05f62e347