[roll] Roll fuchsia [ffx] Add Frame to Playground runtime
A Frame is something between a stack frame and a register set in the
playground interpreter. Loosely it is a container for the values of all
the variables in a particular scope.

Frames do some complicated synchronicity work around capturing, when a
function or closure captures variables from its outer scope. This breaks
the relationship between code flow and variable accesses so we need to
introduce an additional mutex when this happens.

Frames consist of slots which are numbered. The compiler knows which
number matches which variable at a given time so the names can be
discarded. This patch also contains the related data structure for
global variables, where we have to keep the names of variables preserved
so that the next command, run in a different compiler pass, can resolve
them.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987296
Original-Revision: d0571008b851f37bf532ffbd5cef81fc9c603ab6
GitOrigin-RevId: 7796faadd815467c45b67e7b05b8d75de70dec85
Change-Id: I86b4897134c98ca33f9cb98628860668e1cd12b3