[roll] Roll fuchsia [build] Additional fast-math configuration
Add a configuration to enable `-ffast-math` without the
accompanying flags `-fhonor-infinities`, `-fhonor-nans` and
`-fno-finite-math-only`.

In order to do this, a third "baseline" config is created
that sets the common flags, but leaves the finite-math-only
flag to be explicitly set -- or explicitly cleared -- by the
more specific config.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988786
Original-Revision: 54702bdaf5e39f17ab5514140ad146ec9f038571
GitOrigin-RevId: d5187c6e153dda92e0d7b01a37c986b3ef39875c
Change-Id: Ie0079e1f9ff831688c9ad65e6e78b1d52a2ccd20