[roll] Roll fuchsia [loadbench] Switch to using the ProfileResource
This is part of the ProfileProvider deprecation effort. The
loadbench shell package currently uses the ProfileProvider service
to create a bunch of threads with varying deadlines and priorities.
We could switch this over to using the `RoleManager` and define a bunch
of roles for this use case, but that's a very unwieldy setup for a shell
package stress test.

Since this binary is only used in an SSH shell, we can just create
profiles directly with the Profile resource. Note that I did have to
modify the security policy to allow the `ProfileResource` to be routed
to sshd, but seeing as this is a very small subset of the privileges
granted to the sshd component already (e.g. the root job and root
resource), I don't think this increases our attack surface.

Original-Bug: 323262398
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991681
Original-Revision: 0112a4d8390e792f475cd9ff398ea2ddc0fc65c0
GitOrigin-RevId: 4acedd53991cdf5cb14cf480f9a79583e5c78244
Change-Id: Ia4cca212b6a4164697fe91dddd8f15d06fa6503a