[roll] Roll fuchsia [netstack3] Lookup Device to Join Multicast Group
When joining a multicast group via the IP_ADD_MEMBERSHIP (Or
IPV6_ADD_MEMBERSHIP) socket option, callers may provide the device with
which they'd like to join the group. In the case that no selection is
made by the caller, the netstack should select a suitable default device
(i.e. a device with a route to the corresponding multicast address).

Prior to this CL, this case was unhandled, and resulted in callers
observing ENODEV. This CL adds support for this case by consulting the
routing table and selecting an appropriate default device.

Original-Fixed: b/323355372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984652
Original-Revision: 7e51a9c2a98b487a8bc58c2347c24b9cd3f3a9a8
GitOrigin-RevId: 1b4c3f8a364e8536d3dec706b4b43b91195412e6
Change-Id: I2189122758b5d5d04fc8385cb7c6e33c3b0d3179