[roll] Roll fuchsia [zxdb][line_input] Ignore Ctrl-C in embedded mode
This CL teaches zxdb not to handle Ctrl-C (and other similar keystrokes,
like Ctrl-D) in the embedded console mode.

The key is to put the terminal back into "normal" mode, which cases the
pty to intercept these keystrokes and deliver them to the controller of
the pty, which is typically the shell. When zxdb runs inside `fx test`,
this causes a Python exception to be delivered to `fx test`, which is
what it normally expects.

Even though zxdb is responsible for watching stdin for input, the
`line_input` library is responsible for controlling the terminal mode.
For this reason, we plumb the information about whether we want to
listen for these keystrokes through the `line_input` library.

Original-Fixed: b/324926524
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993065
Original-Revision: d15163db142fd21031917b910d00344673a39b19
GitOrigin-RevId: 574385394b8c517b15c4a1446d833310bc42860e
Change-Id: If8d5253e1e4a67901345a7df555ef56c9c9ff937