[roll] Roll fuchsia [kernel][arm64] Refactor iframe setup slightly
This refactors the assembly macros somewhat to reduce false
sharing and make it easier to understand the from-EL0 and
from-EL1 paths independently.  In the EL1 path, it no longer
saves and restores the SP_EL0 value in the iframe's USP slot, but
instead stores the interrupted SP (which is always the same as
`(iframe_t*)iframe + 1`) and ignores it on restore since an EL0
register value doesn't belong in an EL1 iframe but it's useful to
have a uniform "interrupted SP" value to read out in iframe dumps
and such whether for an EL0 or an EL1 iframe.  In the EL0 path,
this shaves off a few redundant instructions to fetch TPIDR_EL1.

Original-Bug: 323463901
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985952
Original-Revision: 511494d943dd3ccd4b7b39dfd70814549b230636
GitOrigin-RevId: 490ab15d1905f3792972c8388bb26cbbe19b9526
Change-Id: Ib58400e0a74b68aec530477dff3fa70fa51b4676