[roll] Roll fuchsia [debugger] Add AnalyticsReporter and Session events
This adds a way for the Session object to provide analytics events to
the rest of the client. Basic events are added for the Session object,
which report what the Session is connected to and how long the session
lasts.

Users of the zxdb client library will have analytics disabled by
default, but may enable them if they so choose.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993059
Original-Revision: e2918419cedf09efc50468c950d6db38b81a95b9
GitOrigin-RevId: f2b37ffca93d97caa0981243de7c0f9817187394
Change-Id: I1d65423b8602a437eed51ef4ba91aac6b5760624