[roll] Roll fuchsia [amlogic-display][dsi] Inject PanelConfig to DsiHost.
This adds the PanelConfig const pointer to DsiHost ctor so that the
PanelConfig will be injected and the reference will be stored
in DsiHost ctor.

The PanelConfig will be converted to a display_setting_t and sent
to the DSI Host controller driver to configure the DSI transmitter.

Test: no behavior change expected.
Test: core.nelson booted.
Original-Bug: 318479388
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969344
Original-Revision: 3d4f00f33b7e3f3a82672e600b3761b02406e85a
GitOrigin-RevId: f2daf1e2e1cc435fad0b46c5abcd1423d21e09e2
Change-Id: I87c206941c0f66ae0c56370646e02e95f7c557f2