[roll] Roll fuchsia [display][fidl] Move fuchsia.hardware.display.types/ImageId.
fuchsia.hardware.display.types/ImageId represents image IDs in the
coordinator-client namespace. Moving it outside of types makes it
clearer that it's not supposed to be used for image IDs in the
coordinator-driver namespace.

This CL also clarifies the comments for the ImageId and DriverImageId
types in the api-types-cpp library.

Original-Bug: 316631158
Test: none (no functional changes)
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993072
Original-Revision: b6c1d1581e26483befdf50399d3ef04144fc4f9f
GitOrigin-RevId: 8630fcc9035e325658df720385134ee713d82bc5
Change-Id: Ib7eacc695bef8b166f84467cd07ce8e41b7c69e0