[roll] Roll fuchsia [component_manager] Make Router async
Update Router::route from a synchronous function that uses
a completer to an async function. This cleans up the places
in the component manager codebase where the component's
task group had to be use to create an async context.

The main reason the change to async is better is because
it is a useful pattern to write a Router that calls
another router to get a capability *and then* operate
on that capability. This is much easier in an async
context.

This code is a bit awkward because Rust doesn't support
async llambdas. A good clean up here would be to replace
some of these llambdas with structs that impl Routable.
Then Router could hold a Box<dyn Routable>.

Original-Bug: b/325096862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/992415
Original-Revision: 39001268300e014868d9e2fb78f6cd1ba0510c55
GitOrigin-RevId: f3db9c7b1dedadab8fb6af3122e6a6abae86efaa
Change-Id: I1e3796d95f675676da68d0b8f064578d0a7b0300