[roll] Roll infra/recipes [run_script] Support shallow checkouts
This recipe is used to check out a repository that contain a lot of
large frequently-modified binaries, so the git history is massive and
growing rapidly, making checkouts take up to 20 minutes (and rising) and
often time out.

We should be able to stem the bleeding by doing a shallow checkout for
this repository.

Original-Bug: 325644661
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/994212
Original-Revision: 23fb1855ea7f87ccebd8d56fb24988fb60176de0
GitOrigin-RevId: 85bf0f1bbeef9c6c10eff06b5ec8fa47f3b510d3
Change-Id: I417706b8ab396d06810d8717afa2b85c176b4fcc