[roll] Roll fuchsia [honeydew] Fix tracing implementation.
The loop attempting to read the channel does not register for updates
when ZX_SHOULD_WAIT is returned from the channel, leading to a flurry of
missed channel updates while the channel is held that are never read
(and lots of log spam). Uses the AsyncChannel implementation instead so
that this is read/downloaded properly.

This also adds additional verification testing to the functional test to
ensure that tracing is actually happening and valid data is being
recorded. So far this only verifies that the trace can be converted into
JSON and that at least one "kernel:syscall" category (a default if no
categories are listed) is included in the trace.

Multiply: tracing_test_fc
Original-Bug: b/324123125
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991220
Original-Revision: 82b00cbece9d20959550df1d1965da6467c03221
GitOrigin-RevId: 45248ecced2b27336106de5372e706a9a8d953c4
Change-Id: I90d7ed6078122df403f2c2d33458d7f4c707e56f