[roll] Roll fuchsia [dsi-dw] Move the driver implementation to a standalone library.
This change moves the dsi-dw driver implementation to a standalone
class (designware_dsi::DsiHostController) in a standalone library.
The DsiHostController class uses the same interface and banjo types
as DsiDw.

In a future change, we'll add a DsiHostController class instance to
the amlogic-display driver so that the amlogic-display driver won't
need to do inter-driver banjo calls.

After that, we'll be able to delete the DsiDw driver safely, and then
we can replace the banjo types used within the amlogic-display driver.

Test: no-op. panel power off/on works on core.nelson.
Original-Bug: 319379442
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993338
Original-Revision: b3173d88db6dee2f08197778a2aadb54a5857904
GitOrigin-RevId: 1f1a9b59d04048381438ac9e8d356f83daefecfa
Change-Id: I472d0ed9377d9c5931205f8c39866ec57ef281f6