[roll] Roll fuchsia [wlansoftmac] Create SendPtr and refactor completer types
In their previous form, the completer types used by the bridged
wlansoftmac relied on casting a pointer as a usize to avoid having to
implement `Send` for a wrapper around them. Now that we wish to use a
`SendExecutor` for the bridged wlansoftmac soon, and will therefore
actually be sending these pointers to another thread, it makes sense
to create a more robust implementation to safely send each completer
between threads.

This refactor implements a new `SendPtr` type to wrap pointers with
when they are safe to send to another thread. As of this change,
`SendPtr` can only be constructed from a `*mut c_void` which is always
safe to send to another thread as long as it always remains a `*mut
c_void`. In particular, the pointee of a `*mut c_void` cannot be read
or mutated because it's a `c_void`.

The long-term plan for `SendPtr` is that it will only implement unsafe
constructor where the caller must make a particular promise about the
pointer being used to construct the `SendPtr`.

This change also makes various improvements to various safety-related
documentation in wlansoftmac-c.

Original-Bug: 42075638
Multiply: fuchsia-pkg://fuchsia.com/wlan-mlme-tests
Multiply: fuchsia-pkg://fuchsia.com/wlansoftmac-rust-tests
Multiply: fuchsia-pkg://fuchsia.com/wlan-hw-sim-contemporary-privacy-tests
Multiply: fuchsia-pkg://fuchsia.com/wlan-hw-sim-legacy-privacy-tests
Multiply: fuchsia-pkg://fuchsia.com/wlan-hw-sim-rate-selection-tests
Test: Successful iface creation, scan, connect, disconnect with iwlwifi on NUC
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989041
Original-Revision: d6f317d96301ebadbecc68ca1ea8faa77fcaa3cb
GitOrigin-RevId: b050fe37ddf45f8b8c68033329226160df277de1
Change-Id: Ic36a55f9d96bdd66901df0fb44272d3d5c00249a