[roll] Roll fuchsia Reland "[realm_builder] Use availability attributes for obsolete AddLocalChild"
This is a reland of commit efabd2b688ce05bb465ffa385e4aa93633602506

Reason for reland: Suppressed deprecation warnings due to uses of the
deprecated LocalComponent class in code that supports use of the class
for SDK consumers targeting API levels prior to its removal.

Though this CL backdates deprecation of LocalComponent to API level 9,
the functions with which it is used were already deprecated so there are
unlikely to be any new deprecation warnings in SDK consumers. The most
likely case would be if someone had suppressed the warnings around the
function calls but not use of the type. This pattern has been used
rarely if at all in petals.

Original change's description:
> [realm_builder] Use availability attributes for obsolete AddLocalChild
>
> Replace the __Fuchsia_API_level__ preprocessor check around the
> LocalComponent class (added in https://fxrev.dev/950019) with a
> `ZX_REMOVED_SINCE` attribute. This avoids the empty class declaration
> and the need for some preprocessor conditions. Some tests and method
> definitions are still excluded in API level 17 and later because they
> use the removed type and do not compile with the new definition of
> LocalComponentKind, respectively.
>
> This change to stable API level 17 is safe because it just changes how a
> type and methods are made unavailable at that level.
>
> Also fix some comments that reference API level 15 instead of 16.
>
> Original-Bug: b/42061169, b/296292544
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/989432

Original-Bug: b/42061169, b/296292544
Test: No warnings in `fx build sdk/lib/sys/component/cpp/testing:cpp_library` at API level 14.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993063
Original-Revision: b998e42a201eec31c70da642387d77d18d6fd676
GitOrigin-RevId: 076f793aa8d2fa604a46eeff579fa7dca7df050f
Change-Id: I555e783f7564f8ae9d36c0e49076d8c22ee0e4c6