[roll] Roll fuchsia [netdevice] Ensure shim unbinds asynchronously
The network device shim is a special case because it uses dispatchers
that are not owned by the driver they belong to. This is to allow
inlining of FIDL calls between the main network device driver and the
network device shim. This could also lead to a situation where when
NetworkDeviceShim is tearing down it would call Unbind and it would be
inlined as well. This could lead to a deadlock as the unbinding code is
not prepared to handle a case where posting a task is inlined. It ended
up waiting for a completion that could only be signaled from the thread
it was blocking.

Solve this by posting the Unbind call in the shim to the dispatcher to
ensure asynchronous behavior. Because the Unbind call now happens from
the same dispatcher that the task is posted to, it cannot be inlined
anymore.

Test: Added unit test that triggers the tricky teardown sequence.
Original-Bug: b/325595277
Multiply: network-tun-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994100
Original-Revision: b3c3ca32a30435063a91a883fd4e03a54dce8783
GitOrigin-RevId: 8a05cf2aa6b284a40f98b4f46c6672d63769a7da
Change-Id: I584f800abb1088fa53bef92506d2694065ddf3ac