[roll] Roll fuchsia [fidl][media] Remove @transitional
The @transitional attribute is for short transitions but there are a few
in the media APIs that have been in progress for many years. All of this
has been replaced by versioning.

This removes @transitional from methods and adds implementations to
servers that panic, in case they were being called in error.

Original-Bug: 42078310
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994272
Original-Revision: db58e38e21df89ba2834793becf833ab18349e66
GitOrigin-RevId: 025d8de2788459f529a45112226d7d9080f1d2d4
Change-Id: I27b6e31e8708889c9a4d2704c407e1799a8f1731