[roll] Roll fuchsia [compat] Lower log severity
This log can occur when stop was triggered before the init hook, but the
driver tries to complete it's init hook once triggered. This sequence is
valid and the driver did nothing wrong, so this shouldn't be logged as
an error.

Original-Fixed: 318983912
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994664
Original-Revision: e738a83e81bca5c88e538f31580ce6516c105804
GitOrigin-RevId: c4c86810be89fa817409caeecc7ed0b6b1cef651
Change-Id: I0e6b413447627e856eb4fe4a4525e6fd73ded1dd