[roll] Roll fuchsia [netdevice] Make tests correctly wait for startup
With the move to FIDL there was a change in how the
NetworkDeviceImpl::Start sequence occurred. Previously the completion of
Start (through callback(cookie, status)) processed the remaining part of
the Start sequence inline. Now the completion of Start (through
completer.Reply(status)) is asynchronous and the effects of the
remaining part of the Start sequence are not necessarily observed just
because reply is called. This could cause problems in tests that relied
on the old behavior. If the remaining Start sequence was delayed enough
it could interfere with other parts of the test by send RX queue trigger
packets at inopportune times.

Fix this by adding a callback in FakeNetworkDeviceImpl::Start that tests
can use to install an RX queue trigger handler before the rest of the
Start sequence is about to run. This ensures that the RX queue trigger
from the Start sequence is consumed before tests proceed and that all
other setup is done as well.

Original-Bug: b/323601448
Multiply: network-device-lib-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986492
Original-Revision: b2343f95c36a8fbea69975dfd90e82e73c4a9784
GitOrigin-RevId: 36d3497ee96c7d46a26c5ab0d488240809abff24
Change-Id: Id04523848873c383b62c6f758d17d0574882f20e