[roll] Roll fuchsia [bazel][sdk] Fail if an API level is not set
If a user does not set an API level when building a fuchsia package with
the bazel sdk we will fail.

We do not fail if the attribute is not set because people can set an api
level on the command line. In the future we might want to revisit this
decision if the errors are not clear.

BUG: b/317417932
BUG: b/303683945
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/993240
Original-Revision: f712f146a2e3b9d8b7381907d650744b26fed1b7
GitOrigin-RevId: fb4c4d7cba44359666958d491671f53c125b9cea
Change-Id: I5e4229121c0a4419205c9e1220d5eae1d6ec045f