[roll] Roll fuchsia [libc] Remove extraneous fence from call_once
The atomic_load on control establishes a happens-before relationship
with the atomic_exchange on control as both use the (implicit) memory
order seq_cst.

The pthread_once implementation was originally based on a bespoke
memory model using volatile load/stores and memory barriers. This
fence was updated to a C11 atomic fence in this change:
https://fuchsia.googlesource.com/fuchsia/+/13e7f0cdbfa9ba53c714180089306816abcd7fea%5E%21/#F3

The other synchronization operations were updated to use C11 atomics
in this later change:
https://fuchsia.googlesource.com/fuchsia/+/f2373edae22bc921a41286318243b3985102b510%5E%21/#F9

But the fence operation was left in place.

This was noticed in our storage benchmarks when boringssl added calls
to CRYPTO_library_init which internally calls pthread_once to many
entry points.

Original-Bug: https://fxbug.dev/324138139

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994159
Original-Revision: 0221b74216992af9a20e076aadc19998833b5356
GitOrigin-RevId: 37ba0367a24674bc0de9038de3b50f0b25cc4232
Change-Id: I792a4376e67a0cb414e6f99a047ed0075630d604