[roll] Roll fuchsia [tests][golden_files] Allowing overriding of test label
This adds a parameter to golden_files template named "visible_label"

Some usages of golden_files do not have global visibility, as a result
the error message printing the label to make sure is included in the
build graph could be not visible.

The visible_label parameter is used to specify a visible label for this
message. If it is not given, the label of the golden_file test is used
as normal.

Original-Bug: b/326258617
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994961
Original-Revision: f4733a8515509ffb6ab07576ec24c63620cf7c8d
GitOrigin-RevId: 524bed0a8942658ac8b81179cd812f62f97b59b8
Change-Id: Ie997b1811e90ec85aab25a8bbccaaa5695d31c27