[roll] Roll fuchsia [filter.deprecated] Add FIDL for probing filter
Add a method to the Filter protocol to check whether there is a server
attached to the other end. This will be used by netcfg to check whether
the optional capability it has is actually being served before using it
to create clients. If the optional capability is present, then netcfg
will use the deprecated protocol instead of the new filter protocol.
The bug to track the usage of that API is https://fxbug.dev/323950209

Original-Bug: 323950209
Original-Fixed: 323949472
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987244
Original-Revision: d512a8713d08455ada92605a8a316c2ac6d97fbe
GitOrigin-RevId: 5bd778a7e7b3b2efe1fa91c2c9bacc2ed9843564
Change-Id: I6a2ffa51268148063527adf8977fe9774943ed92