[roll] Roll fuchsia [trace_manager] Use async::Executor
As part of an effort to allow better reasoning about async control flow
in trace_manger and reducing blocking on the async loop, it will be
helpful to add the ability to rely more heavily on the Fuchsia async
promise libraries.

In this patch, we lift the `async_dispatcher_t*` held by the various
objects in trace_manager into an `async::Executor&`. When
async::Executors are destroyed they shutdown the async loop, so we need
to keep a single global one in `main` and pass around references to it.

Original-Bug: 295183613
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995083
Original-Revision: 8e1d8fb08132471ecf8a2bd6a96af13c4349b565
GitOrigin-RevId: 519c39deef04f1b3686125312b0061d924b732be
Change-Id: I04dab8aca4807154e33e4f0d566fc3d909e9acdd