[roll] Roll infra/recipes [LLDB] Use real prefix for Python
This leaks strings from our build into the built Python, but it allows
pip and setuptools to be installed, fixing an issue in the LLDB test
suite. This appears to be an intrinsic limitation of the Python build;
it takes cues from the prefix to install pip and setuptools as part of
the `make install` process, so my hacked prefix of `/.` doesn't fly.

Original-Bug: 326077194
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/994761
Original-Revision: 7941f14ae1ce03f4699d4dd9a9db811c861107a6
GitOrigin-RevId: c16168264e61f7f2771a0124088fb4ddf7e58885
Change-Id: I712f673201d39fcfe4d38eaf5e2f17cf554b215e