[roll] Roll fuchsia [kernel][phys] Add ArchPatchInfo type and plumbing
This adds the ArchPatchInfo type in physboot.  This will collect
any state that physboot should calculate once and then might be
needed be ArchPatchCode and/or ArchHandoff.  It's separate from
ArchPhysInfo because this is for data specific to physboot's
patching of the kernel proper and any arch-specific handoff logic
that may interrelate with that data, while ArchPhysInfo is for
things that the low-level startup of any phys executable will do.

Original-Bug: 322202704
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995832
Original-Revision: 3a7571d52b6acf983ff069f1da412f02e5640d18
GitOrigin-RevId: af6e23173a4ffee4771a3346d5abe9862a4c6899
Change-Id: I60ce6825f3197beef58542275be0dc65ce64e2bf