[roll] Roll fuchsia [bazel_sdk][tests] bazel_test.py: Incrementality fix.
The script tries to create a symlink to a specific target file,
but the logic used failed when an existing dangling symlink already
existed at the current destination location, even though the
original code tried to deal (incorrectly) with this situation,
due to the subtle differences between os.path.exists(),
os.path.lexists() and pathlib.Path.exists().

To fix this, implement a `force_symlink()` that always removes
the destination link in case of EEXIST. This is similar to what
is done in other Python scripts of our build system (but
bazel_test.py cannot rely on them, since it is supposed to be
usable out-of-tree).

Original-Bug: 326567616
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996052
Original-Revision: 07d87e08c7b7fe9f8d67a504642a80480fdc302b
GitOrigin-RevId: bbb0768fff1fe519a190597167cd9ba4ad78aeb6
Change-Id: I8a0628344ea6fcd464b06acf040e4d41ded24c1b