[roll] Roll fuchsia [netdevice] Keep test events alive during calls
The network device library has a few events that are used for testing
purposes. These events are represented by fit::functions. A recent
change to fix a flake started using one of these events to ensure that
tests were correctly waiting for the network device start sequence to
finish. The fix included replacing an event handler during the
invocation of the same event handler. This introduced a problem where
because of this the old event handler, currently executing, was
destroyed before it completed. The event handler could then attempt to
access destroyed state. In some instances this lead to a captured this
pointer being null, even if the object it pointed to was still alive.

Fix this by wrapping each call to these event handlers in a function
that first shares the event handler and maintains a local reference for
the duration of the invocation. This way the event handler can safely be
replaced without being destroyed until the invocation of the event
handler has finished.

Test: Verified that this type of behavior does not destruct captures.
Original-Bug: b/323601448
Multiply: network-device-lib-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995507
Original-Revision: 03585c792a2dce3f508dade21dab645cb99978ae
GitOrigin-RevId: 4aee8b87471028c5388ff90760fc7e812a9c8c3b
Change-Id: Idf1472a600023729747631320518c3200f3b9474