[roll] Roll fuchsia [wlan][policy] Fix e2e test dependency on async ordering
After receiving a client state update, the tests did not _ensure_ that
the ack was processed. This CL fixes that, by running the main future
after the ack is sent. Prior to http://fxrev.dev/990573, the timing was
different, and the ack was processed as a side effect of some other
`run_while()` in the test.

In making this change, another minor ordering assumption was revealed in
`prepare_client_interface()`, which previously did not ensure that the
state machine fully processed its initial setup before the end of the
`prepare_client_interface()` function.

Test: fx test wlancfg-tests
Original-Bug: 324986229
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995393
Original-Revision: 780a8b4708334fd5b7f339cb109c150e60e8321a
GitOrigin-RevId: 899ad63788493299b7dcb04ecba339e4eb4e6a61
Change-Id: I36b7e9c3346461a1f6e476e2dfd1842d7c25c8ce