[roll] Roll fuchsia [debugger] Print the real filter pattern in attach
When using process names, we trim the string to zircon process name max
length, which modifies the string. We print a warning that we are doing
this, but then print the unmodified string we were given, which is
confusing.

Make sure to update the pattern we print (which is also used for things
like job koids) to the trimmed string.

Original-Bug: b/193336765
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995744
Original-Revision: bb8f00f2342b328b72a6522d2a0cd166a7f102e0
GitOrigin-RevId: e8ff9de93b45f493b4e2ca034b9f38e8624fa4d2
Change-Id: Ib261b8e22a8463ee3ea9517d8db8904301a6e190