[roll] Roll fuchsia [ddk-environment-test] Delete test
This test made more sense in DFv1 when we manually launching the DFv1
driver host and providing it with it's incoming namespace. Now that we
are not doing that, this test would simply serve to validate CF is
actually providing the protocols specified in the driver host's
component manifest, which isn't very useful.

Original-Bug: 316176095
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995733
Original-Revision: 5f5bdd67cd2b8a30e522481deb0d2eac2f15ec30
GitOrigin-RevId: 9c00c00711b12c8503503a18214cf3b58fba6548
Change-Id: I7aab501a833f3a52639da2c4486689787a6d3d35