[roll] Roll fuchsia [netstack3] Put datagram sockets behind references
Change the datagram module to expect datagram sockets to be behind
references and propagate the changes to Udp, Icmp sockets.

A follow-up change will move the bindings HashMap on socket IDs to
opaque data in core.

This follows the exact same model as TCP, including the lock ordering
layout. Core keeps a set of all the sockets it knows of, which is used
for inspect (not yet implemented for datagram) and nursing of Primary
references.

Original-Bug: 42076297

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/994097
Original-Revision: 14ef7e31916eb224fe39927596ee815504882979
GitOrigin-RevId: 46af4b9cd58d4fffba463e7b91b2fbdafdbc369e
Change-Id: I1aa859e637c6c0523bffa563b027d94151b09014