[roll] Roll fuchsia [archivist] Fix issue that caused NS3 data to be unavailable in pipelines
Without this fix we would remove the NS3 static selectors when we
handled the diagnostics dir closed event. The diagnostics dir and the
InspectSink handle were treated as coming from different components,
even if they were the same, due to the difference in the URL with which
they were launched.

An alternative fix would be to ensure that the URL hardcoded in the
Archivist is the same as the URL in netstack CMLs. However, I think this
change is correct regardless, as we typically care about the moniker
only, not the URL with which the component was launched at that
location. That's just metadata.

Multiply: archivist-integration-tests
Multiply: archivist-unittests
Original-Fixed: 326510415
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995813
Original-Revision: 0e9fed516794120c92d391f9b3e65ce129800979
GitOrigin-RevId: 71cea496c212eac4738d3b9e3706b84b3562a3fc
Change-Id: I58e70c4dd19985634754a58b90c6603967f925ee