[roll] Roll fuchsia [logs] Add support for transactional symbolizer
This CL adds support for the transactional symbolizer
protocol to the log_formatter's definition of what a symbolizer is.
Future patches will implement the log_formatter changes
for handling transactions and the transactional symbolizer itself.
This merely defines the interface to the transactional symbolizer.
The new method exists to facilitate a soft transition to the
transactional symbolizer so it can be enabled or disabled
at runtime with a flag. If the symbolzer is not transactional,
all logging works the same as it has before. For log_listener,
symbolize is a no-op so it can safely claim to support the transactional
path. The old symbolizer won't support transactions so returns
false to indicate to log_formatter that it is invalid to
attempt transactions on it.

Original-Bug: b/299980894
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/996363
Original-Revision: 65283895afb7475f49878c8c1fc5ced939252953
GitOrigin-RevId: 2c124760df411f9ce562f7988d41e002b345f292
Change-Id: Ied76f6b9bed82d189dcc8a0247e339d1611b2ad8