[roll] Roll fuchsia [kernel][vm] Move aspace query in PageFault
Move call to QueryAspaceAndReplace so that it is checked earlier, which
saves the unnecessary call to IfExistsPages in the case that a page was
already mapped to the address. In this case, the pages retrieved from
IfExistsPages are not utilized.

Original-Bug: 42085825
Run-All-Tests: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/990574
Original-Revision: a618ec5b19995ac10a9031f5867ab2c94c926924
GitOrigin-RevId: 9378136ac360b48280e7a1446282f610d76d1195
Change-Id: I6f739ade41cf1f75cb2b8ec06e3b9c78a9f140c4