[roll] Roll fuchsia [kernel] Count wasted preemption events
Add two new kcounters.  One is incremented whenever the scheduler sets
the preemption timer to a point that's prior to the time latched when
the CPU last entered the scheduler.  The other is incremented when the
scheduler dequeues a deadline task with a finish_time that's prior to
the eligible time (now).  See https://fxbug.dev/42182770 for more
context.

Original-Bug: 42182770
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/995494
Original-Revision: 117f09c7f553332d122ca8dc73d0eaa27b343860
GitOrigin-RevId: ac171a4767c3887b19d38222244f344dfa894ca7
Change-Id: I2d00aa633a3dee7cd5036009612189a85660c6e1